@@ -1,68 +1,70 @@
 {
-  "name": "TheWildWind",
-  "version": "1.0.0",
-  "main": "server/server.js",
-  "engines": {
-    "node": "9.11.1"
-  },
-  "scripts": {
-    "lint": "eslint .",
-    "start": "node server/server.js",
-    "posttest": "npm run lint && nsp check"
-  },
-  "dependencies": {
-    "axios": "^0.18.0",
-    "compression": "^1.0.3",
-    "cors": "^2.5.2",
-    "helmet": "^3.10.0",
-    "history": "^4.7.2",
-    "loopback": "^3.19.0",
-    "loopback-boot": "^2.6.5",
-    "loopback-component-explorer": "^5.4.0",
-    "loopback-connector-mongodb": "^3.4.3",
-    "npm": "^5.8.0",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
-    "react-loadable": "^5.4.0",
-    "react-redux": "^5.0.7",
-    "react-router-dom": "^4.2.2",
-    "react-router-redux": "^4.0.8",
-    "redux": "^3.7.2",
-    "redux-devtools-extension": "^2.13.2",
-    "redux-thunk": "^2.2.0",
-    "serve-favicon": "^2.0.1",
-    "strong-error-handler": "^2.0.0",
-    "superagent": "^3.8.3",
-    "webpack": "^4.8.0",
-    "webpack-cli": "^2.1.3"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
-    "css-loader": "^0.28.11",
-    "cssnano": "^3.10.0",
-    "eslint": "^3.17.1",
-    "eslint-config-loopback": "^8.0.0",
-    "file-loader": "^1.1.11",
-    "nsp": "^2.1.0",
-    "postcss-cssnext": "^3.1.0",
-    "postcss-loader": "^2.1.5",
-    "prop-types": "^15.6.1",
-    "semantic-ui-react": "^0.80.0",
-    "style-loader": "^0.20.3",
-    "url-loader": "^1.0.1",
-    "webpack-bundle-analyzer": "^2.11.1"
-  },
-  "repository": {
-    "type": "",
-    "url": ""
-  },
-  "license": "UNLICENSED",
-  "description": "TheWildWind"
+    "name": "TheWildWind",
+    "version": "1.0.0",
+    "main": "server/server.js",
+    "engines": {
+        "node": "9.11.1"
+    },
+    "scripts": {
+        "lint": "eslint .",
+        "start": "node server/server.js",
+        "posttest": "npm run lint && nsp check"
+    },
+    "dependencies": {
+        "axios": "^0.18.0",
+        "compression": "^1.0.3",
+        "cors": "^2.5.2",
+        "helmet": "^3.10.0",
+        "history": "^4.7.2",
+        "loopback": "^3.19.0",
+        "loopback-boot": "^2.6.5",
+        "loopback-component-explorer": "^5.4.0",
+        "loopback-connector-mongodb": "^3.4.3",
+        "npm": "^5.8.0",
+        "react": "^16.3.2",
+        "react-dom": "^16.3.2",
+        "react-loadable": "^5.4.0",
+        "react-redux": "^5.0.7",
+        "react-router-dom": "^4.2.2",
+        "react-router-redux": "^4.0.8",
+        "redux": "^3.7.2",
+        "redux-devtools-extension": "^2.13.2",
+        "redux-thunk": "^2.2.0",
+        "serve-favicon": "^2.0.1",
+        "strong-error-handler": "^2.0.0",
+        "superagent": "^3.8.3",
+        "webpack": "^4.8.0",
+        "webpack-cli": "^2.1.3"
+    },
+    "devDependencies": {
+        "babel-cli": "^6.26.0",
+        "babel-core": "^6.26.3",
+        "babel-loader": "^7.1.4",
+        "babel-plugin-syntax-dynamic-import": "^6.18.0",
+        "babel-plugin-transform-class-properties": "^6.24.1",
+        "babel-plugin-transform-object-rest-spread": "^6.26.0",
+        "babel-preset-env": "^1.6.1",
+        "babel-preset-react": "^6.24.1",
+        "babel-eslint": "^8.2.3",
+        "eslint-plugin-react": "^7.8.2",
+        "css-loader": "^0.28.11",
+        "cssnano": "^3.10.0",
+        "eslint": "^4.19.0",
+        "eslint-config-loopback": "^8.0.0",
+        "file-loader": "^1.1.11",
+        "nsp": "^2.1.0",
+        "postcss-cssnext": "^3.1.0",
+        "postcss-loader": "^2.1.5",
+        "prop-types": "^15.6.1",
+        "semantic-ui-react": "^0.80.0",
+        "style-loader": "^0.20.3",
+        "url-loader": "^1.0.1",
+        "webpack-bundle-analyzer": "^2.11.1"
+    },
+    "repository": {
+        "type": "",
+        "url": ""
+    },
+    "license": "UNLICENSED",
+    "description": "TheWildWind"
 }
