@@ -42,7 +42,7 @@ eval("\nvar content = __webpack_require__(/*! !../../../node_modules/css-loader!
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _Image = __webpack_require__(/*! semantic-ui-react/dist/commonjs/elements/Image */ \"./node_modules/semantic-ui-react/dist/commonjs/elements/Image/index.js\");\n\nvar _Image2 = _interopRequireDefault(_Image);\n\nvar _Button = __webpack_require__(/*! semantic-ui-react/dist/commonjs/elements/Button */ \"./node_modules/semantic-ui-react/dist/commonjs/elements/Button/index.js\");\n\nvar _Button2 = _interopRequireDefault(_Button);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\nvar _redux = __webpack_require__(/*! redux */ \"./node_modules/redux/es/index.js\");\n\nvar _itemActions = __webpack_require__(/*! actions/itemActions */ \"./src/node_modules/actions/itemActions.js\");\n\n__webpack_require__(/*! ./catalog.css */ \"./src/node_modules/catalog/catalog.css\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Catalog = function (_Component) {\n    _inherits(Catalog, _Component);\n\n    function Catalog(props) {\n        _classCallCheck(this, Catalog);\n\n        var _this = _possibleConstructorReturn(this, (Catalog.__proto__ || Object.getPrototypeOf(Catalog)).call(this, props));\n\n        _this.state = {\n            items: [],\n            showModal: false\n        };\n        return _this;\n    }\n\n    _createClass(Catalog, [{\n        key: 'componentDidMount',\n        value: function componentDidMount() {\n            this.props.get();\n        }\n    }, {\n        key: 'componentWillReceiveProps',\n        value: function componentWillReceiveProps(_ref) {\n            var itemList = _ref.itemList,\n                view = _ref.view,\n                addToBasket = _ref.addToBasket;\n\n            if (itemList !== this.state.items) {\n                var items = itemList.map(function (item) {\n                    return _react2.default.createElement(\n                        'div',\n                        { className: 'item-box', key: item.id },\n                        _react2.default.createElement(\n                            'div',\n                            { className: 'item-grid' },\n                            _react2.default.createElement(\n                                'div',\n                                { className: 'item-desc-header' },\n                                item.name\n                            ),\n                            _react2.default.createElement(\n                                'div',\n                                { className: 'item-image' },\n                                _react2.default.createElement(_Image2.default, { src: item.image })\n                            ),\n                            _react2.default.createElement(\n                                'div',\n                                { className: 'item-desc' },\n                                _react2.default.createElement('p', { className: 'item-desc-text' })\n                            ),\n                            _react2.default.createElement(\n                                'div',\n                                { className: 'item-buttons' },\n                                _react2.default.createElement(\n                                    _Button2.default.Group,\n                                    { vertical: true, floated: 'right' },\n                                    _react2.default.createElement(_Button2.default, {\n                                        content: '\\u041F\\u043E\\u0434\\u0440\\u043E\\u0431\\u043D\\u0435\\u0435',\n                                        icon: 'search',\n                                        labelPosition: 'left',\n                                        onClick: function onClick() {\n                                            return view(item.id);\n                                        }\n                                    }),\n                                    _react2.default.createElement(_Button2.default, {\n                                        content: '\\u0414\\u043E\\u0431\\u0430\\u0432\\u0438\\u0442\\u044C',\n                                        icon: 'shop',\n                                        labelPosition: 'left',\n                                        onClick: function onClick() {\n                                            return addToBasket(item, 1);\n                                        }\n                                    })\n                                )\n                            )\n                        )\n                    );\n                });\n                this.setState({\n                    items: items\n                });\n            }\n        }\n    }, {\n        key: 'render',\n        value: function render() {\n            return _react2.default.createElement(\n                'div',\n                { className: 'item-container' },\n                this.state.items\n            );\n        }\n    }]);\n\n    return Catalog;\n}(_react.Component);\n\nvar mapStateToProps = function mapStateToProps(state) {\n    return {\n        itemList: state.items.list,\n        showModal: state.items.showModal\n    };\n};\n\nvar mapDispatchToProps = function mapDispatchToProps(dispatch) {\n    return (0, _redux.bindActionCreators)({\n        get: _itemActions.getItems,\n        view: _itemActions.viewItem,\n        addToBasket: _itemActions.addToBasket\n    }, dispatch);\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps, mapDispatchToProps)(Catalog);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2NhdGFsb2cvaW5kZXguanMuanMiLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vc3JjL25vZGVfbW9kdWxlcy9jYXRhbG9nL2luZGV4LmpzP2I4NDUiXSwic291cmNlc0NvbnRlbnQiOlsiaW1wb3J0IFJlYWN0LCB7Q29tcG9uZW50fSBmcm9tICdyZWFjdCdcbmltcG9ydCBJbWFnZSBmcm9tICdzZW1hbnRpYy11aS1yZWFjdC9kaXN0L2NvbW1vbmpzL2VsZW1lbnRzL0ltYWdlJ1xuaW1wb3J0IEJ1dHRvbiBmcm9tICdzZW1hbnRpYy11aS1yZWFjdC9kaXN0L2NvbW1vbmpzL2VsZW1lbnRzL0J1dHRvbidcbmltcG9ydCB7Y29ubmVjdH0gZnJvbSAncmVhY3QtcmVkdXgnXG5pbXBvcnQge2JpbmRBY3Rpb25DcmVhdG9yc30gZnJvbSAncmVkdXgnXG5pbXBvcnQge2dldEl0ZW1zLCB2aWV3SXRlbSwgYWRkVG9CYXNrZXR9IGZyb20gJ2FjdGlvbnMvaXRlbUFjdGlvbnMnXG5pbXBvcnQgJy4vY2F0YWxvZy5jc3MnXG5cblxuY2xhc3MgQ2F0YWxvZyBleHRlbmRzIENvbXBvbmVudCB7XG4gICAgY29uc3RydWN0b3IocHJvcHMpIHtcbiAgICAgICAgc3VwZXIocHJvcHMpXG4gICAgICAgIHRoaXMuc3RhdGUgPSB7XG4gICAgICAgICAgICBpdGVtczogW10sXG4gICAgICAgICAgICBzaG93TW9kYWw6IGZhbHNlLFxuICAgICAgICB9XG4gICAgfVxuICAgIGNvbXBvbmVudERpZE1vdW50KCkge1xuICAgICAgICB0aGlzLnByb3BzLmdldCgpXG4gICAgfVxuICAgIGNvbXBvbmVudFdpbGxSZWNlaXZlUHJvcHMoe2l0ZW1MaXN0LHZpZXcsYWRkVG9CYXNrZXR9KSB7XG4gICAgICAgIGlmIChpdGVtTGlzdCAhPT0gdGhpcy5zdGF0ZS5pdGVtcykge1xuICAgICAgICAgICAgY29uc3QgaXRlbXMgPSBpdGVtTGlzdC5tYXAoKGl0ZW0pID0+IHtcbiAgICAgICAgICAgICAgICByZXR1cm4gKFxuICAgICAgICAgICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cIml0ZW0tYm94XCIga2V5PXtpdGVtLmlkfT5cbiAgICAgICAgICAgICAgICAgICAgICAgIDxkaXYgY2xhc3NOYW1lPVwiaXRlbS1ncmlkXCI+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJpdGVtLWRlc2MtaGVhZGVyXCI+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIHtpdGVtLm5hbWV9XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJpdGVtLWltYWdlXCI+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIDxJbWFnZSBzcmM9e2l0ZW0uaW1hZ2V9Lz5cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cIml0ZW0tZGVzY1wiPlxuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICA8cCBjbGFzc05hbWU9XCJpdGVtLWRlc2MtdGV4dFwiPlxuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICA8L3A+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJpdGVtLWJ1dHRvbnNcIj5cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgPEJ1dHRvbi5Hcm91cCB2ZXJ0aWNhbCBmbG9hdGVkPSdyaWdodCc+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICA8QnV0dG9uXG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgY29udGVudD0n0J/QvtC00YDQvtCx0L3QtdC1J1xuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIGljb249J3NlYXJjaCdcbiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICBsYWJlbFBvc2l0aW9uPSdsZWZ0J1xuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIG9uQ2xpY2s9eygpID0+IHZpZXcoaXRlbS5pZCl9XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAvPlxuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgPEJ1dHRvblxuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIGNvbnRlbnQ9J9CU0L7QsdCw0LLQuNGC0YwnXG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgaWNvbj0nc2hvcCdcbiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICBsYWJlbFBvc2l0aW9uPSdsZWZ0J1xuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIG9uQ2xpY2s9eygpID0+IGFkZFRvQmFza2V0KGl0ZW0sMSl9XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAvPlxuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICA8L0J1dHRvbi5Hcm91cD5cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgICAgICAgICAgICAgICAgIDwvZGl2PlxuICAgICAgICAgICAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgICAgICAgICApXG4gICAgICAgICAgICB9KVxuICAgICAgICAgICAgdGhpcy5zZXRTdGF0ZSh7XG4gICAgICAgICAgICAgICAgaXRlbXM6IGl0ZW1zXG4gICAgICAgICAgICB9KVxuICAgICAgICB9XG4gICAgfVxuICAgIHJlbmRlcigpIHtcbiAgICAgICAgcmV0dXJuIChcbiAgICAgICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cIml0ZW0tY29udGFpbmVyXCI+XG4gICAgICAgICAgICAgICAgICAgIHt0aGlzLnN0YXRlLml0ZW1zfVxuICAgICAgICAgICAgICAgIDwvZGl2PlxuICAgICAgICApXG4gICAgfVxufVxuXG5jb25zdCBtYXBTdGF0ZVRvUHJvcHMgPSBzdGF0ZSA9PiB7XG4gICAgcmV0dXJuIHtcbiAgICAgICAgaXRlbUxpc3Q6IHN0YXRlLml0ZW1zLmxpc3QsXG4gICAgICAgIHNob3dNb2RhbDogc3RhdGUuaXRlbXMuc2hvd01vZGFsXG4gICAgfVxufVxuXG5jb25zdCBtYXBEaXNwYXRjaFRvUHJvcHMgPSBkaXNwYXRjaCA9PiB7XG4gICAgcmV0dXJuIGJpbmRBY3Rpb25DcmVhdG9ycyh7XG4gICAgICAgIGdldDogZ2V0SXRlbXMsXG4gICAgICAgIHZpZXc6IHZpZXdJdGVtLFxuICAgICAgICBhZGRUb0Jhc2tldDogYWRkVG9CYXNrZXQsXG4gICAgfSwgZGlzcGF0Y2gpXG59XG5cblxuXG5leHBvcnQgZGVmYXVsdFxuY29ubmVjdChtYXBTdGF0ZVRvUHJvcHMsbWFwRGlzcGF0Y2hUb1Byb3BzKShDYXRhbG9nKVxuIl0sIm1hcHBpbmdzIjoiOzs7Ozs7OztBQUFBO0FBQ0E7OztBQUFBO0FBQ0E7OztBQUFBO0FBQ0E7OztBQUFBO0FBQ0E7QUFBQTtBQUNBO0FBQUE7QUFDQTtBQUFBO0FBQ0E7Ozs7Ozs7OztBQUVBOzs7QUFDQTtBQUFBO0FBQ0E7QUFEQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBRkE7QUFGQTtBQU1BO0FBQ0E7OztBQUFBO0FBQ0E7QUFDQTs7O0FBQ0E7QUFBQTtBQUFBO0FBQUE7QUFDQTtBQUFBO0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFDQTtBQUFBO0FBQUE7QUFDQTtBQUFBO0FBQUE7QUFDQTtBQURBO0FBR0E7QUFBQTtBQUFBO0FBQ0E7QUFEQTtBQUdBO0FBQUE7QUFBQTtBQUNBO0FBREE7QUFJQTtBQUFBO0FBQUE7QUFDQTtBQUFBO0FBQUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFBQTtBQUpBO0FBTUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUFBO0FBQUE7QUFKQTtBQVBBO0FBREE7QUFYQTtBQURBO0FBK0JBO0FBQ0E7QUFDQTtBQURBO0FBR0E7QUFDQTs7O0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFDQTtBQURBO0FBSUE7Ozs7QUExREE7QUFDQTtBQTREQTtBQUNBO0FBQ0E7QUFDQTtBQUZBO0FBSUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFIQTtBQUtBO0FBQ0E7QUFJQSIsInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./src/node_modules/catalog/index.js\n");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _Image = __webpack_require__(/*! semantic-ui-react/dist/commonjs/elements/Image */ \"./node_modules/semantic-ui-react/dist/commonjs/elements/Image/index.js\");\n\nvar _Image2 = _interopRequireDefault(_Image);\n\nvar _Button = __webpack_require__(/*! semantic-ui-react/dist/commonjs/elements/Button */ \"./node_modules/semantic-ui-react/dist/commonjs/elements/Button/index.js\");\n\nvar _Button2 = _interopRequireDefault(_Button);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\nvar _redux = __webpack_require__(/*! redux */ \"./node_modules/redux/es/index.js\");\n\nvar _itemActions = __webpack_require__(/*! actions/itemActions */ \"./src/node_modules/actions/itemActions.js\");\n\n__webpack_require__(/*! ./catalog.css */ \"./src/node_modules/catalog/catalog.css\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Catalog = function (_Component) {\n    _inherits(Catalog, _Component);\n\n    function Catalog(props) {\n        _classCallCheck(this, Catalog);\n\n        var _this = _possibleConstructorReturn(this, (Catalog.__proto__ || Object.getPrototypeOf(Catalog)).call(this, props));\n\n        _this.state = {\n            items: [],\n            showModal: false\n        };\n        return _this;\n    }\n\n    _createClass(Catalog, [{\n        key: 'componentDidMount',\n        value: function componentDidMount() {\n            this.props.get();\n        }\n    }, {\n        key: 'componentWillReceiveProps',\n        value: function componentWillReceiveProps(_ref) {\n            var itemList = _ref.itemList,\n                view = _ref.view,\n                addToBasket = _ref.addToBasket;\n\n            if (itemList !== this.state.items) {\n                var items = itemList.map(function (item) {\n                    return _react2.default.createElement(\n                        'div',\n                        { className: 'item-box', key: item.id },\n                        _react2.default.createElement(\n                            'div',\n                            { className: 'item-grid' },\n                            _react2.default.createElement(\n                                'div',\n                                { className: 'item-desc-header' },\n                                item.name\n                            ),\n                            _react2.default.createElement(\n                                'div',\n                                { className: 'item-image' },\n                                _react2.default.createElement(_Image2.default, { src: item.image })\n                            ),\n                            _react2.default.createElement(\n                                'div',\n                                { className: 'item-desc' },\n                                _react2.default.createElement('p', { className: 'item-desc-text' })\n                            ),\n                            _react2.default.createElement(\n                                'div',\n                                { className: 'item-buttons' },\n                                _react2.default.createElement(\n                                    _Button2.default.Group,\n                                    { vertical: true, floated: 'right' },\n                                    _react2.default.createElement(_Button2.default, {\n                                        content: '\\u041F\\u043E\\u0434\\u0440\\u043E\\u0431\\u043D\\u0435\\u0435',\n                                        icon: 'search',\n                                        labelPosition: 'left',\n                                        onClick: function onClick() {\n                                            return view(item.id);\n                                        }\n                                    }),\n                                    _react2.default.createElement(_Button2.default, {\n                                        content: '\\u0414\\u043E\\u0431\\u0430\\u0432\\u0438\\u0442\\u044C',\n                                        icon: 'shop',\n                                        labelPosition: 'left',\n                                        onClick: function onClick() {\n                                            return addToBasket(item, 1);\n                                        }\n                                    })\n                                )\n                            )\n                        )\n                    );\n                });\n                this.setState({\n                    items: items\n                });\n            }\n        }\n    }, {\n        key: 'render',\n        value: function render() {\n            return _react2.default.createElement(\n                'div',\n                { className: 'item-container' },\n                this.state.items\n            );\n        }\n    }]);\n\n    return Catalog;\n}(_react.Component);\n\nvar mapStateToProps = function mapStateToProps(state) {\n    return {\n        itemList: state.items.list,\n        showModal: state.items.showModal\n    };\n};\n\nvar mapDispatchToProps = function mapDispatchToProps(dispatch) {\n    return (0, _redux.bindActionCreators)({\n        get: _itemActions.getItems,\n        view: _itemActions.viewItem,\n        addToBasket: _itemActions.addToBasket\n    }, dispatch);\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps, mapDispatchToProps)(Catalog);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2NhdGFsb2cvaW5kZXguanMuanMiLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vc3JjL25vZGVfbW9kdWxlcy9jYXRhbG9nL2luZGV4LmpzP2I4NDUiXSwic291cmNlc0NvbnRlbnQiOlsiaW1wb3J0IFJlYWN0LCB7Q29tcG9uZW50fSBmcm9tICdyZWFjdCdcbmltcG9ydCBJbWFnZSBmcm9tICdzZW1hbnRpYy11aS1yZWFjdC9kaXN0L2NvbW1vbmpzL2VsZW1lbnRzL0ltYWdlJ1xuaW1wb3J0IEJ1dHRvbiBmcm9tICdzZW1hbnRpYy11aS1yZWFjdC9kaXN0L2NvbW1vbmpzL2VsZW1lbnRzL0J1dHRvbidcbmltcG9ydCB7Y29ubmVjdH0gZnJvbSAncmVhY3QtcmVkdXgnXG5pbXBvcnQge2JpbmRBY3Rpb25DcmVhdG9yc30gZnJvbSAncmVkdXgnXG5pbXBvcnQge2dldEl0ZW1zLCB2aWV3SXRlbSwgYWRkVG9CYXNrZXR9IGZyb20gJ2FjdGlvbnMvaXRlbUFjdGlvbnMnXG5pbXBvcnQgJy4vY2F0YWxvZy5jc3MnXG5cblxuY2xhc3MgQ2F0YWxvZyBleHRlbmRzIENvbXBvbmVudCB7XG4gICAgY29uc3RydWN0b3IocHJvcHMpIHtcbiAgICAgICAgc3VwZXIocHJvcHMpXG4gICAgICAgIHRoaXMuc3RhdGUgPSB7XG4gICAgICAgICAgICBpdGVtczogW10sXG4gICAgICAgICAgICBzaG93TW9kYWw6IGZhbHNlLFxuICAgICAgICB9XG4gICAgfVxuICAgIGNvbXBvbmVudERpZE1vdW50KCkge1xuICAgICAgICB0aGlzLnByb3BzLmdldCgpXG4gICAgfVxuICAgIGNvbXBvbmVudFdpbGxSZWNlaXZlUHJvcHMoe2l0ZW1MaXN0LHZpZXcsYWRkVG9CYXNrZXR9KSB7XG4gICAgICAgIGlmIChpdGVtTGlzdCAhPT0gdGhpcy5zdGF0ZS5pdGVtcykge1xuICAgICAgICAgICAgY29uc3QgaXRlbXMgPSBpdGVtTGlzdC5tYXAoKGl0ZW0pID0+IHtcbiAgICAgICAgICAgICAgICByZXR1cm4gKFxuICAgICAgICAgICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cIml0ZW0tYm94XCIga2V5PXtpdGVtLmlkfT5cbiAgICAgICAgICAgICAgICAgICAgICAgIDxkaXYgY2xhc3NOYW1lPVwiaXRlbS1ncmlkXCI+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJpdGVtLWRlc2MtaGVhZGVyXCI+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIHtpdGVtLm5hbWV9XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJpdGVtLWltYWdlXCI+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIDxJbWFnZSBzcmM9e2l0ZW0uaW1hZ2V9Lz5cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cIml0ZW0tZGVzY1wiPlxuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICA8cCBjbGFzc05hbWU9XCJpdGVtLWRlc2MtdGV4dFwiPlxuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICA8L3A+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJpdGVtLWJ1dHRvbnNcIj5cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgPEJ1dHRvbi5Hcm91cCB2ZXJ0aWNhbCBmbG9hdGVkPSdyaWdodCc+XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICA8QnV0dG9uXG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgY29udGVudD0n0J/QvtC00YDQvtCx0L3QtdC1J1xuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIGljb249J3NlYXJjaCdcbiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICBsYWJlbFBvc2l0aW9uPSdsZWZ0J1xuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIG9uQ2xpY2s9eygpID0+IHZpZXcoaXRlbS5pZCl9XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAvPlxuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgPEJ1dHRvblxuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIGNvbnRlbnQ9J9CU0L7QsdCw0LLQuNGC0YwnXG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgaWNvbj0nc2hvcCdcbiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICBsYWJlbFBvc2l0aW9uPSdsZWZ0J1xuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIG9uQ2xpY2s9eygpID0+IGFkZFRvQmFza2V0KGl0ZW0sMSl9XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAvPlxuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICA8L0J1dHRvbi5Hcm91cD5cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgICAgICAgICAgICAgICAgIDwvZGl2PlxuICAgICAgICAgICAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgICAgICAgICApXG4gICAgICAgICAgICB9KVxuICAgICAgICAgICAgdGhpcy5zZXRTdGF0ZSh7XG4gICAgICAgICAgICAgICAgaXRlbXM6IGl0ZW1zXG4gICAgICAgICAgICB9KVxuICAgICAgICB9XG4gICAgfVxuICAgIHJlbmRlcigpIHtcbiAgICAgICAgcmV0dXJuIChcbiAgICAgICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cIml0ZW0tY29udGFpbmVyXCI+XG4gICAgICAgICAgICAgICAgICAgIHt0aGlzLnN0YXRlLml0ZW1zfVxuICAgICAgICAgICAgICAgIDwvZGl2PlxuICAgICAgICApXG4gICAgfVxufVxuXG5jb25zdCBtYXBTdGF0ZVRvUHJvcHMgPSBzdGF0ZSA9PiB7XG4gICAgcmV0dXJuIHtcbiAgICAgICAgaXRlbUxpc3Q6IHN0YXRlLml0ZW1zLmxpc3QsXG4gICAgICAgIHNob3dNb2RhbDogc3RhdGUuaXRlbXMuc2hvd01vZGFsXG4gICAgfVxufVxuXG5jb25zdCBtYXBEaXNwYXRjaFRvUHJvcHMgPSBkaXNwYXRjaCA9PiB7XG4gICAgcmV0dXJuIGJpbmRBY3Rpb25DcmVhdG9ycyh7XG4gICAgICAgIGdldDogZ2V0SXRlbXMsXG4gICAgICAgIHZpZXc6IHZpZXdJdGVtLFxuICAgICAgICBhZGRUb0Jhc2tldDogYWRkVG9CYXNrZXQsXG4gICAgfSwgZGlzcGF0Y2gpXG59XG5cblxuXG5leHBvcnQgZGVmYXVsdFxuY29ubmVjdChtYXBTdGF0ZVRvUHJvcHMsbWFwRGlzcGF0Y2hUb1Byb3BzKShDYXRhbG9nKVxuIl0sIm1hcHBpbmdzIjoiOzs7Ozs7OztBQUFBO0FBQ0E7OztBQUFBO0FBQ0E7OztBQUFBO0FBQ0E7OztBQUFBO0FBQ0E7QUFBQTtBQUNBO0FBQUE7QUFDQTtBQUFBO0FBQ0E7Ozs7Ozs7OztBQUVBOzs7QUFDQTtBQUFBO0FBQ0E7QUFEQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBRkE7QUFGQTtBQU1BO0FBQ0E7OztBQUFBO0FBQ0E7QUFDQTs7O0FBQ0E7QUFBQTtBQUFBO0FBQUE7QUFDQTtBQUFBO0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFDQTtBQUFBO0FBQUE7QUFDQTtBQUFBO0FBQUE7QUFDQTtBQURBO0FBR0E7QUFBQTtBQUFBO0FBQ0E7QUFEQTtBQUdBO0FBQUE7QUFBQTtBQUNBO0FBREE7QUFJQTtBQUFBO0FBQUE7QUFDQTtBQUFBO0FBQUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFBQTtBQUpBO0FBTUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUFBO0FBQUE7QUFKQTtBQVBBO0FBREE7QUFYQTtBQURBO0FBK0JBO0FBQ0E7QUFDQTtBQURBO0FBR0E7QUFDQTs7O0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFDQTtBQURBO0FBSUE7Ozs7OztBQUdBO0FBQ0E7QUFDQTtBQUNBO0FBRkE7QUFJQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUhBO0FBS0E7QUFDQTtBQUlBIiwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./src/node_modules/catalog/index.js\n");
 
 /***/ })
 
