@@ -1,5 +1,16 @@
 (window["webpackJsonp"] = window["webpackJsonp"] || []).push([[4],{
 
+/***/ "./node_modules/css-loader/index.js!./node_modules/postcss-loader/lib/index.js!./src/node_modules/basket/basket.css":
+/*!********************************************************************************************************!*\
+  !*** ./node_modules/css-loader!./node_modules/postcss-loader/lib!./src/node_modules/basket/basket.css ***!
+  \********************************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports, __webpack_require__) {
+
+eval("exports = module.exports = __webpack_require__(/*! ../../../node_modules/css-loader/lib/css-base.js */ \"./node_modules/css-loader/lib/css-base.js\")(false);\n// imports\n\n\n// module\nexports.push([module.i, \".basket{display:-webkit-box;display:-ms-flexbox;display:flex;-webkit-box-orient:vertical;-webkit-box-direction:normal;-ms-flex-direction:column;flex-direction:column;-webkit-box-flex:2;-ms-flex:2 0 auto;flex:2 0 auto;margin:16px 32px;margin:1rem 2rem;-webkit-box-pack:justify;-ms-flex-pack:justify;justify-content:space-between}.basket-checkout{justify-self:flex-end}.label{margin-bottom:0;margin-top:.5em}\", \"\"]);\n\n// exports\n//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9ub2RlX21vZHVsZXMvY3NzLWxvYWRlci9pbmRleC5qcyEuL25vZGVfbW9kdWxlcy9wb3N0Y3NzLWxvYWRlci9saWIvaW5kZXguanMhLi9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9iYXNrZXQuY3NzLmpzIiwic291cmNlcyI6WyJ3ZWJwYWNrOi8vLy4vc3JjL25vZGVfbW9kdWxlcy9iYXNrZXQvYmFza2V0LmNzcz8yN2FjIl0sInNvdXJjZXNDb250ZW50IjpbImV4cG9ydHMgPSBtb2R1bGUuZXhwb3J0cyA9IHJlcXVpcmUoXCIuLi8uLi8uLi9ub2RlX21vZHVsZXMvY3NzLWxvYWRlci9saWIvY3NzLWJhc2UuanNcIikoZmFsc2UpO1xuLy8gaW1wb3J0c1xuXG5cbi8vIG1vZHVsZVxuZXhwb3J0cy5wdXNoKFttb2R1bGUuaWQsIFwiLmJhc2tldHtkaXNwbGF5Oi13ZWJraXQtYm94O2Rpc3BsYXk6LW1zLWZsZXhib3g7ZGlzcGxheTpmbGV4Oy13ZWJraXQtYm94LW9yaWVudDp2ZXJ0aWNhbDstd2Via2l0LWJveC1kaXJlY3Rpb246bm9ybWFsOy1tcy1mbGV4LWRpcmVjdGlvbjpjb2x1bW47ZmxleC1kaXJlY3Rpb246Y29sdW1uOy13ZWJraXQtYm94LWZsZXg6MjstbXMtZmxleDoyIDAgYXV0bztmbGV4OjIgMCBhdXRvO21hcmdpbjoxNnB4IDMycHg7bWFyZ2luOjFyZW0gMnJlbTstd2Via2l0LWJveC1wYWNrOmp1c3RpZnk7LW1zLWZsZXgtcGFjazpqdXN0aWZ5O2p1c3RpZnktY29udGVudDpzcGFjZS1iZXR3ZWVufS5iYXNrZXQtY2hlY2tvdXR7anVzdGlmeS1zZWxmOmZsZXgtZW5kfS5sYWJlbHttYXJnaW4tYm90dG9tOjA7bWFyZ2luLXRvcDouNWVtfVwiLCBcIlwiXSk7XG5cbi8vIGV4cG9ydHNcbiJdLCJtYXBwaW5ncyI6IkFBQUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTsiLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./node_modules/css-loader/index.js!./node_modules/postcss-loader/lib/index.js!./src/node_modules/basket/basket.css\n");
+
+/***/ }),
+
 /***/ "./node_modules/css-loader/index.js!./node_modules/postcss-loader/lib/index.js!./src/node_modules/components/dropdown/dropdown.css":
 /*!***********************************************************************************************************************!*\
   !*** ./node_modules/css-loader!./node_modules/postcss-loader/lib!./src/node_modules/components/dropdown/dropdown.css ***!
@@ -7,7 +18,18 @@
 /*! no static exports found */
 /***/ (function(module, exports, __webpack_require__) {
 
-eval("exports = module.exports = __webpack_require__(/*! ../../../../node_modules/css-loader/lib/css-base.js */ \"./node_modules/css-loader/lib/css-base.js\")(false);\n// imports\n\n\n// module\nexports.push([module.i, \".dd-wrapper{position:relative;width:320px;width:20rem}.dd-header{margin-bottom:0}.dd-header,.dd-list{border:1px solid #000;border-radius:2%}.dd-list{list-style-type:none;padding:0;margin-top:0}.dd-list-item{border-bottom:1px solid #000;background-color:rgba(200,255,200,.3)}.dd-list-item:last-child{border-bottom:none}.dd-list-item.selected{background-color:rgba(200,255,200,.7)}\", \"\"]);\n\n// exports\n//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9ub2RlX21vZHVsZXMvY3NzLWxvYWRlci9pbmRleC5qcyEuL25vZGVfbW9kdWxlcy9wb3N0Y3NzLWxvYWRlci9saWIvaW5kZXguanMhLi9zcmMvbm9kZV9tb2R1bGVzL2NvbXBvbmVudHMvZHJvcGRvd24vZHJvcGRvd24uY3NzLmpzIiwic291cmNlcyI6WyJ3ZWJwYWNrOi8vLy4vc3JjL25vZGVfbW9kdWxlcy9jb21wb25lbnRzL2Ryb3Bkb3duL2Ryb3Bkb3duLmNzcz9mMmZmIl0sInNvdXJjZXNDb250ZW50IjpbImV4cG9ydHMgPSBtb2R1bGUuZXhwb3J0cyA9IHJlcXVpcmUoXCIuLi8uLi8uLi8uLi9ub2RlX21vZHVsZXMvY3NzLWxvYWRlci9saWIvY3NzLWJhc2UuanNcIikoZmFsc2UpO1xuLy8gaW1wb3J0c1xuXG5cbi8vIG1vZHVsZVxuZXhwb3J0cy5wdXNoKFttb2R1bGUuaWQsIFwiLmRkLXdyYXBwZXJ7cG9zaXRpb246cmVsYXRpdmU7d2lkdGg6MzIwcHg7d2lkdGg6MjByZW19LmRkLWhlYWRlcnttYXJnaW4tYm90dG9tOjB9LmRkLWhlYWRlciwuZGQtbGlzdHtib3JkZXI6MXB4IHNvbGlkICMwMDA7Ym9yZGVyLXJhZGl1czoyJX0uZGQtbGlzdHtsaXN0LXN0eWxlLXR5cGU6bm9uZTtwYWRkaW5nOjA7bWFyZ2luLXRvcDowfS5kZC1saXN0LWl0ZW17Ym9yZGVyLWJvdHRvbToxcHggc29saWQgIzAwMDtiYWNrZ3JvdW5kLWNvbG9yOnJnYmEoMjAwLDI1NSwyMDAsLjMpfS5kZC1saXN0LWl0ZW06bGFzdC1jaGlsZHtib3JkZXItYm90dG9tOm5vbmV9LmRkLWxpc3QtaXRlbS5zZWxlY3RlZHtiYWNrZ3JvdW5kLWNvbG9yOnJnYmEoMjAwLDI1NSwyMDAsLjcpfVwiLCBcIlwiXSk7XG5cbi8vIGV4cG9ydHNcbiJdLCJtYXBwaW5ncyI6IkFBQUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTsiLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./node_modules/css-loader/index.js!./node_modules/postcss-loader/lib/index.js!./src/node_modules/components/dropdown/dropdown.css\n");
+eval("exports = module.exports = __webpack_require__(/*! ../../../../node_modules/css-loader/lib/css-base.js */ \"./node_modules/css-loader/lib/css-base.js\")(false);\n// imports\n\n\n// module\nexports.push([module.i, \".dd-wrapper{position:relative;width:320px;width:20rem}.dd-header{margin-bottom:0}.dd-header,.dd-list{border:1px solid #000;border-radius:2%;background-color:#fff}.dd-list{list-style-type:none;padding:0;margin-top:0}.dd-list .dd-list-item{border-bottom:1px solid #000;background-color:rgba(200,255,200,.3)}.dd-list-item:last-child{border-bottom:none}.dd-list-item.selected{background-color:rgba(200,255,200,.7)}\", \"\"]);\n\n// exports\n//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9ub2RlX21vZHVsZXMvY3NzLWxvYWRlci9pbmRleC5qcyEuL25vZGVfbW9kdWxlcy9wb3N0Y3NzLWxvYWRlci9saWIvaW5kZXguanMhLi9zcmMvbm9kZV9tb2R1bGVzL2NvbXBvbmVudHMvZHJvcGRvd24vZHJvcGRvd24uY3NzLmpzIiwic291cmNlcyI6WyJ3ZWJwYWNrOi8vLy4vc3JjL25vZGVfbW9kdWxlcy9jb21wb25lbnRzL2Ryb3Bkb3duL2Ryb3Bkb3duLmNzcz9mMmZmIl0sInNvdXJjZXNDb250ZW50IjpbImV4cG9ydHMgPSBtb2R1bGUuZXhwb3J0cyA9IHJlcXVpcmUoXCIuLi8uLi8uLi8uLi9ub2RlX21vZHVsZXMvY3NzLWxvYWRlci9saWIvY3NzLWJhc2UuanNcIikoZmFsc2UpO1xuLy8gaW1wb3J0c1xuXG5cbi8vIG1vZHVsZVxuZXhwb3J0cy5wdXNoKFttb2R1bGUuaWQsIFwiLmRkLXdyYXBwZXJ7cG9zaXRpb246cmVsYXRpdmU7d2lkdGg6MzIwcHg7d2lkdGg6MjByZW19LmRkLWhlYWRlcnttYXJnaW4tYm90dG9tOjB9LmRkLWhlYWRlciwuZGQtbGlzdHtib3JkZXI6MXB4IHNvbGlkICMwMDA7Ym9yZGVyLXJhZGl1czoyJTtiYWNrZ3JvdW5kLWNvbG9yOiNmZmZ9LmRkLWxpc3R7bGlzdC1zdHlsZS10eXBlOm5vbmU7cGFkZGluZzowO21hcmdpbi10b3A6MH0uZGQtbGlzdCAuZGQtbGlzdC1pdGVte2JvcmRlci1ib3R0b206MXB4IHNvbGlkICMwMDA7YmFja2dyb3VuZC1jb2xvcjpyZ2JhKDIwMCwyNTUsMjAwLC4zKX0uZGQtbGlzdC1pdGVtOmxhc3QtY2hpbGR7Ym9yZGVyLWJvdHRvbTpub25lfS5kZC1saXN0LWl0ZW0uc2VsZWN0ZWR7YmFja2dyb3VuZC1jb2xvcjpyZ2JhKDIwMCwyNTUsMjAwLC43KX1cIiwgXCJcIl0pO1xuXG4vLyBleHBvcnRzXG4iXSwibWFwcGluZ3MiOiJBQUFBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7Iiwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./node_modules/css-loader/index.js!./node_modules/postcss-loader/lib/index.js!./src/node_modules/components/dropdown/dropdown.css\n");
+
+/***/ }),
+
+/***/ "./src/node_modules/basket/basket.css":
+/*!********************************************!*\
+  !*** ./src/node_modules/basket/basket.css ***!
+  \********************************************/
+/*! no static exports found */
+/***/ (function(module, exports, __webpack_require__) {
+
+eval("\nvar content = __webpack_require__(/*! !../../../node_modules/css-loader!../../../node_modules/postcss-loader/lib!./basket.css */ \"./node_modules/css-loader/index.js!./node_modules/postcss-loader/lib/index.js!./src/node_modules/basket/basket.css\");\n\nif(typeof content === 'string') content = [[module.i, content, '']];\n\nvar transform;\nvar insertInto;\n\n\n\nvar options = {\"hmr\":true}\n\noptions.transform = transform\noptions.insertInto = undefined;\n\nvar update = __webpack_require__(/*! ../../../node_modules/style-loader/lib/addStyles.js */ \"./node_modules/style-loader/lib/addStyles.js\")(content, options);\n\nif(content.locals) module.exports = content.locals;\n\nif(false) {}//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9iYXNrZXQuY3NzLmpzIiwic291cmNlcyI6WyJ3ZWJwYWNrOi8vLy4vc3JjL25vZGVfbW9kdWxlcy9iYXNrZXQvYmFza2V0LmNzcz9lMzhkIl0sInNvdXJjZXNDb250ZW50IjpbIlxudmFyIGNvbnRlbnQgPSByZXF1aXJlKFwiISEuLi8uLi8uLi9ub2RlX21vZHVsZXMvY3NzLWxvYWRlci9pbmRleC5qcyEuLi8uLi8uLi9ub2RlX21vZHVsZXMvcG9zdGNzcy1sb2FkZXIvbGliL2luZGV4LmpzIS4vYmFza2V0LmNzc1wiKTtcblxuaWYodHlwZW9mIGNvbnRlbnQgPT09ICdzdHJpbmcnKSBjb250ZW50ID0gW1ttb2R1bGUuaWQsIGNvbnRlbnQsICcnXV07XG5cbnZhciB0cmFuc2Zvcm07XG52YXIgaW5zZXJ0SW50bztcblxuXG5cbnZhciBvcHRpb25zID0ge1wiaG1yXCI6dHJ1ZX1cblxub3B0aW9ucy50cmFuc2Zvcm0gPSB0cmFuc2Zvcm1cbm9wdGlvbnMuaW5zZXJ0SW50byA9IHVuZGVmaW5lZDtcblxudmFyIHVwZGF0ZSA9IHJlcXVpcmUoXCIhLi4vLi4vLi4vbm9kZV9tb2R1bGVzL3N0eWxlLWxvYWRlci9saWIvYWRkU3R5bGVzLmpzXCIpKGNvbnRlbnQsIG9wdGlvbnMpO1xuXG5pZihjb250ZW50LmxvY2FscykgbW9kdWxlLmV4cG9ydHMgPSBjb250ZW50LmxvY2FscztcblxuaWYobW9kdWxlLmhvdCkge1xuXHRtb2R1bGUuaG90LmFjY2VwdChcIiEhLi4vLi4vLi4vbm9kZV9tb2R1bGVzL2Nzcy1sb2FkZXIvaW5kZXguanMhLi4vLi4vLi4vbm9kZV9tb2R1bGVzL3Bvc3Rjc3MtbG9hZGVyL2xpYi9pbmRleC5qcyEuL2Jhc2tldC5jc3NcIiwgZnVuY3Rpb24oKSB7XG5cdFx0dmFyIG5ld0NvbnRlbnQgPSByZXF1aXJlKFwiISEuLi8uLi8uLi9ub2RlX21vZHVsZXMvY3NzLWxvYWRlci9pbmRleC5qcyEuLi8uLi8uLi9ub2RlX21vZHVsZXMvcG9zdGNzcy1sb2FkZXIvbGliL2luZGV4LmpzIS4vYmFza2V0LmNzc1wiKTtcblxuXHRcdGlmKHR5cGVvZiBuZXdDb250ZW50ID09PSAnc3RyaW5nJykgbmV3Q29udGVudCA9IFtbbW9kdWxlLmlkLCBuZXdDb250ZW50LCAnJ11dO1xuXG5cdFx0dmFyIGxvY2FscyA9IChmdW5jdGlvbihhLCBiKSB7XG5cdFx0XHR2YXIga2V5LCBpZHggPSAwO1xuXG5cdFx0XHRmb3Ioa2V5IGluIGEpIHtcblx0XHRcdFx0aWYoIWIgfHwgYVtrZXldICE9PSBiW2tleV0pIHJldHVybiBmYWxzZTtcblx0XHRcdFx0aWR4Kys7XG5cdFx0XHR9XG5cblx0XHRcdGZvcihrZXkgaW4gYikgaWR4LS07XG5cblx0XHRcdHJldHVybiBpZHggPT09IDA7XG5cdFx0fShjb250ZW50LmxvY2FscywgbmV3Q29udGVudC5sb2NhbHMpKTtcblxuXHRcdGlmKCFsb2NhbHMpIHRocm93IG5ldyBFcnJvcignQWJvcnRpbmcgQ1NTIEhNUiBkdWUgdG8gY2hhbmdlZCBjc3MtbW9kdWxlcyBsb2NhbHMuJyk7XG5cblx0XHR1cGRhdGUobmV3Q29udGVudCk7XG5cdH0pO1xuXG5cdG1vZHVsZS5ob3QuZGlzcG9zZShmdW5jdGlvbigpIHsgdXBkYXRlKCk7IH0pO1xufSJdLCJtYXBwaW5ncyI6IkFBQUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQSIsInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./src/node_modules/basket/basket.css\n");
 
 /***/ }),
 
@@ -19,7 +41,7 @@ eval("exports = module.exports = __webpack_require__(/*! ../../../../node_module
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\nvar _redux = __webpack_require__(/*! redux */ \"./node_modules/redux/es/index.js\");\n\nvar _itemActions = __webpack_require__(/*! actions/itemActions */ \"./src/node_modules/actions/itemActions.js\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar CheckoutButtons = function (_Component) {\n    _inherits(CheckoutButtons, _Component);\n\n    function CheckoutButtons() {\n        _classCallCheck(this, CheckoutButtons);\n\n        return _possibleConstructorReturn(this, (CheckoutButtons.__proto__ || Object.getPrototypeOf(CheckoutButtons)).apply(this, arguments));\n    }\n\n    _createClass(CheckoutButtons, [{\n        key: 'render',\n        value: function render() {\n            var _this2 = this;\n\n            return _react2.default.createElement(\n                'div',\n                { className: 'basket-checkout' },\n                _react2.default.createElement(\n                    'button',\n                    { onClick: function onClick() {\n                            return _this2.props.checkout();\n                        } },\n                    'Checkout'\n                )\n            );\n        }\n    }]);\n\n    return CheckoutButtons;\n}(_react.Component);\n\nvar mapStateToProps = function mapStateToProps(state) {\n    return {\n        basket: state.items.basket\n    };\n};\nvar mapDispatchToProps = function mapDispatchToProps(dispatch) {\n    return (0, _redux.bindActionCreators)({\n        checkout: _itemActions.checkout\n    }, dispatch);\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps, mapDispatchToProps)(CheckoutButtons);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9jaGVja291dC5qcy5qcyIsInNvdXJjZXMiOlsid2VicGFjazovLy9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9jaGVja291dC5qcz8xZGMyIl0sInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBSZWFjdCwge0NvbXBvbmVudH0gZnJvbSAncmVhY3QnXG5pbXBvcnQge2Nvbm5lY3R9IGZyb20gJ3JlYWN0LXJlZHV4J1xuaW1wb3J0IHtiaW5kQWN0aW9uQ3JlYXRvcnN9IGZyb20gJ3JlZHV4J1xuaW1wb3J0IHtjaGVja291dH0gZnJvbSAnYWN0aW9ucy9pdGVtQWN0aW9ucydcblxuY2xhc3MgQ2hlY2tvdXRCdXR0b25zIGV4dGVuZHMgQ29tcG9uZW50IHtcbiAgICByZW5kZXIoKSB7XG4gICAgICAgIHJldHVybihcbiAgICAgICAgICAgIDxkaXYgY2xhc3NOYW1lPVwiYmFza2V0LWNoZWNrb3V0XCI+XG4gICAgICAgICAgICAgICAgPGJ1dHRvbiBvbkNsaWNrPXsoKSA9PiB0aGlzLnByb3BzLmNoZWNrb3V0KCl9PlxuICAgICAgICAgICAgICAgICAgICBDaGVja291dFxuICAgICAgICAgICAgICAgIDwvYnV0dG9uPlxuICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgIClcbiAgICB9XG59XG5cbmNvbnN0IG1hcFN0YXRlVG9Qcm9wcyA9IChzdGF0ZSkgPT4ge1xuICAgIHJldHVybiB7XG4gICAgICAgIGJhc2tldDogc3RhdGUuaXRlbXMuYmFza2V0XG4gICAgfVxufVxuY29uc3QgbWFwRGlzcGF0Y2hUb1Byb3BzID0gZGlzcGF0Y2ggPT4ge1xuICAgIHJldHVybiBiaW5kQWN0aW9uQ3JlYXRvcnMoe1xuICAgICAgICBjaGVja291dDogY2hlY2tvdXRcbiAgICB9LGRpc3BhdGNoKVxufVxuXG5leHBvcnQgZGVmYXVsdCBjb25uZWN0KG1hcFN0YXRlVG9Qcm9wcyxtYXBEaXNwYXRjaFRvUHJvcHMpKENoZWNrb3V0QnV0dG9ucylcbiJdLCJtYXBwaW5ncyI6Ijs7Ozs7Ozs7QUFBQTtBQUNBOzs7QUFBQTtBQUNBO0FBQUE7QUFDQTtBQUFBO0FBQ0E7Ozs7Ozs7OztBQUNBOzs7Ozs7Ozs7OztBQUNBO0FBQUE7QUFDQTtBQUFBO0FBQ0E7QUFBQTtBQUNBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBREE7QUFNQTs7OztBQVRBO0FBQ0E7QUFXQTtBQUNBO0FBQ0E7QUFEQTtBQUdBO0FBQ0E7QUFDQTtBQUNBO0FBREE7QUFHQTtBQUNBO0FBQ0EiLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./src/node_modules/basket/checkout.js\n");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\nvar _redux = __webpack_require__(/*! redux */ \"./node_modules/redux/es/index.js\");\n\nvar _itemActions = __webpack_require__(/*! actions/itemActions */ \"./src/node_modules/actions/itemActions.js\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar CheckoutButtons = function (_Component) {\n    _inherits(CheckoutButtons, _Component);\n\n    function CheckoutButtons(props) {\n        _classCallCheck(this, CheckoutButtons);\n\n        var _this = _possibleConstructorReturn(this, (CheckoutButtons.__proto__ || Object.getPrototypeOf(CheckoutButtons)).call(this, props));\n\n        _this.switchPage = function (next) {\n            return _this.props.switchPage(next);\n        };\n\n        _this.state = {\n            page: 1,\n            basket: _this.props.basket\n        };\n        return _this;\n    }\n\n    _createClass(CheckoutButtons, [{\n        key: 'render',\n        value: function render() {\n            var _this2 = this;\n\n            var switchPage = this.switchPage;\n\n            var next = _react2.default.createElement(\n                'button',\n                { onClick: function onClick() {\n                        return _this2.switchPage(true);\n                    } },\n                '\\u0414\\u0430\\u043B\\u044C\\u0448\\u0435'\n            );\n            var back = _react2.default.createElement(\n                'button',\n                { onClick: function onClick() {\n                        return _this2.switchPage(false);\n                    } },\n                '\\u041D\\u0430\\u0437\\u0430\\u0434'\n            );\n            var checkout = _react2.default.createElement(\n                'button',\n                { onClick: function onClick() {\n                        return _this2.props.checkout();\n                    } },\n                'Checkout'\n            );\n            return _react2.default.createElement(\n                'div',\n                { className: 'basket-checkout' },\n                this.state.page > 1 && back,\n                this.state.page < 3 && next,\n                this.state.page == 3 && checkout\n            );\n        }\n    }], [{\n        key: 'getDerivedStateFromProps',\n        value: function getDerivedStateFromProps(props, state) {\n            if (props.page !== state.page) return {\n                page: props.page\n            };\n        }\n    }]);\n\n    return CheckoutButtons;\n}(_react.Component);\n\nvar mapStateToProps = function mapStateToProps(state) {\n    return {\n        basket: state.items.basket\n    };\n};\nvar mapDispatchToProps = function mapDispatchToProps(dispatch) {\n    return (0, _redux.bindActionCreators)({\n        checkout: _itemActions.checkout\n    }, dispatch);\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps, mapDispatchToProps)(CheckoutButtons);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9jaGVja291dC5qcy5qcyIsInNvdXJjZXMiOlsid2VicGFjazovLy9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9jaGVja291dC5qcz8xZGMyIl0sInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBSZWFjdCwgeyBDb21wb25lbnQgfSBmcm9tICdyZWFjdCdcclxuaW1wb3J0IHsgY29ubmVjdCB9IGZyb20gJ3JlYWN0LXJlZHV4J1xyXG5pbXBvcnQgeyBiaW5kQWN0aW9uQ3JlYXRvcnMgfSBmcm9tICdyZWR1eCdcclxuaW1wb3J0IHsgY2hlY2tvdXQgfSBmcm9tICdhY3Rpb25zL2l0ZW1BY3Rpb25zJ1xyXG5cclxuY2xhc3MgQ2hlY2tvdXRCdXR0b25zIGV4dGVuZHMgQ29tcG9uZW50IHtcclxuICAgIGNvbnN0cnVjdG9yKHByb3BzKSB7XHJcbiAgICAgICAgc3VwZXIocHJvcHMpXHJcbiAgICAgICAgdGhpcy5zdGF0ZSA9IHtcclxuICAgICAgICAgICAgcGFnZTogMSxcclxuICAgICAgICAgICAgYmFza2V0OiB0aGlzLnByb3BzLmJhc2tldFxyXG4gICAgICAgIH1cclxuICAgIH1cclxuICAgIHN0YXRpYyBnZXREZXJpdmVkU3RhdGVGcm9tUHJvcHMocHJvcHMsIHN0YXRlKSB7XHJcbiAgICAgICAgaWYgKHByb3BzLnBhZ2UgIT09IHN0YXRlLnBhZ2UpXHJcbiAgICAgICAgICAgIHJldHVybiB7XHJcbiAgICAgICAgICAgICAgICBwYWdlOiBwcm9wcy5wYWdlXHJcbiAgICAgICAgICAgIH1cclxuICAgIH1cclxuICAgIHN3aXRjaFBhZ2UgPSBuZXh0ID0+IHRoaXMucHJvcHMuc3dpdGNoUGFnZShuZXh0KVxyXG4gICAgcmVuZGVyKCkge1xyXG4gICAgICAgIGNvbnN0IHsgc3dpdGNoUGFnZSB9ID0gdGhpc1xyXG4gICAgICAgIGNvbnN0IG5leHQgPSAoXHJcbiAgICAgICAgICAgIDxidXR0b24gb25DbGljaz17KCkgPT4gdGhpcy5zd2l0Y2hQYWdlKHRydWUpfT7QlNCw0LvRjNGI0LU8L2J1dHRvbj5cclxuICAgICAgICApXHJcbiAgICAgICAgY29uc3QgYmFjayA9IChcclxuICAgICAgICAgICAgPGJ1dHRvbiBvbkNsaWNrPXsoKSA9PiB0aGlzLnN3aXRjaFBhZ2UoZmFsc2UpfT7QndCw0LfQsNC0PC9idXR0b24+XHJcbiAgICAgICAgKVxyXG4gICAgICAgIGNvbnN0IGNoZWNrb3V0ID0gKFxyXG4gICAgICAgICAgICA8YnV0dG9uIG9uQ2xpY2s9eygpID0+IHRoaXMucHJvcHMuY2hlY2tvdXQoKX0+Q2hlY2tvdXQ8L2J1dHRvbj5cclxuICAgICAgICApXHJcbiAgICAgICAgcmV0dXJuIChcclxuICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJiYXNrZXQtY2hlY2tvdXRcIj5cclxuICAgICAgICAgICAgICAgIHt0aGlzLnN0YXRlLnBhZ2UgPiAxICYmIGJhY2t9XHJcbiAgICAgICAgICAgICAgICB7dGhpcy5zdGF0ZS5wYWdlIDwgMyAmJiBuZXh0fVxyXG4gICAgICAgICAgICAgICAge3RoaXMuc3RhdGUucGFnZSA9PSAzICYmIGNoZWNrb3V0fVxyXG4gICAgICAgICAgICA8L2Rpdj5cclxuICAgICAgICApXHJcbiAgICB9XHJcbn1cclxuXHJcbmNvbnN0IG1hcFN0YXRlVG9Qcm9wcyA9IHN0YXRlID0+IHtcclxuICAgIHJldHVybiB7XHJcbiAgICAgICAgYmFza2V0OiBzdGF0ZS5pdGVtcy5iYXNrZXRcclxuICAgIH1cclxufVxyXG5jb25zdCBtYXBEaXNwYXRjaFRvUHJvcHMgPSBkaXNwYXRjaCA9PiB7XHJcbiAgICByZXR1cm4gYmluZEFjdGlvbkNyZWF0b3JzKFxyXG4gICAgICAgIHtcclxuICAgICAgICAgICAgY2hlY2tvdXQ6IGNoZWNrb3V0XHJcbiAgICAgICAgfSxcclxuICAgICAgICBkaXNwYXRjaFxyXG4gICAgKVxyXG59XHJcblxyXG5leHBvcnQgZGVmYXVsdCBjb25uZWN0KG1hcFN0YXRlVG9Qcm9wcywgbWFwRGlzcGF0Y2hUb1Byb3BzKShDaGVja291dEJ1dHRvbnMpXHJcbiJdLCJtYXBwaW5ncyI6Ijs7Ozs7Ozs7QUFBQTtBQUNBOzs7QUFBQTtBQUNBO0FBQUE7QUFDQTtBQUFBO0FBQ0E7Ozs7Ozs7OztBQUNBOzs7QUFDQTtBQUFBO0FBQ0E7QUFEQTtBQUNBO0FBREE7QUFhQTtBQUFBO0FBQ0E7QUFaQTtBQUNBO0FBQ0E7QUFGQTtBQUZBO0FBTUE7QUFDQTs7O0FBT0E7QUFBQTtBQUNBO0FBREE7QUFDQTtBQUNBO0FBQ0E7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBRUE7QUFDQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFFQTtBQUNBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUVBO0FBQ0E7QUFBQTtBQUNBO0FBQ0E7QUFDQTtBQUhBO0FBTUE7OztBQXpCQTtBQUNBO0FBRUE7QUFEQTtBQUdBOzs7Ozs7QUF1QkE7QUFDQTtBQUNBO0FBREE7QUFHQTtBQUNBO0FBQ0E7QUFFQTtBQURBO0FBS0E7QUFDQTtBQUNBIiwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./src/node_modules/basket/checkout.js\n");
 
 /***/ }),
 
@@ -31,7 +53,7 @@ eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\nvar _list = __webpack_require__(/*! ./list */ \"./src/node_modules/basket/list.js\");\n\nvar _list2 = _interopRequireDefault(_list);\n\nvar _checkout = __webpack_require__(/*! ./checkout */ \"./src/node_modules/basket/checkout.js\");\n\nvar _checkout2 = _interopRequireDefault(_checkout);\n\nvar _shipping = __webpack_require__(/*! basket/shipping */ \"./src/node_modules/basket/shipping.js\");\n\nvar _shipping2 = _interopRequireDefault(_shipping);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Basket = function (_Component) {\n    _inherits(Basket, _Component);\n\n    function Basket() {\n        _classCallCheck(this, Basket);\n\n        return _possibleConstructorReturn(this, (Basket.__proto__ || Object.getPrototypeOf(Basket)).apply(this, arguments));\n    }\n\n    _createClass(Basket, [{\n        key: 'render',\n        value: function render() {\n            return this.props.basket.length > 0 ? _react2.default.createElement(\n                'div',\n                { className: 'basket' },\n                _react2.default.createElement(_list2.default, null),\n                _react2.default.createElement(_checkout2.default, null),\n                _react2.default.createElement(_shipping2.default, null)\n            ) : 'Ваша корзина пуста :(';\n        }\n    }]);\n\n    return Basket;\n}(_react.Component);\n\nvar mapStateToProps = function mapStateToProps(state) {\n    return {\n        basket: state.items.basket\n    };\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps)(Basket);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9pbmRleC5qcy5qcyIsInNvdXJjZXMiOlsid2VicGFjazovLy9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9pbmRleC5qcz85ZWNhIl0sInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBSZWFjdCwgeyBDb21wb25lbnQgfSBmcm9tICdyZWFjdCdcbmltcG9ydCB7IGNvbm5lY3QgfSBmcm9tICdyZWFjdC1yZWR1eCdcbmltcG9ydCBCYXNrZXRMaXN0IGZyb20gJy4vbGlzdCdcbmltcG9ydCBDaGVja291dEJ1dHRvbnMgZnJvbSAnLi9jaGVja291dCdcbmltcG9ydCBCYXNrZXRTaGlwcGluZyBmcm9tICdiYXNrZXQvc2hpcHBpbmcnXG5cbmNsYXNzIEJhc2tldCBleHRlbmRzIENvbXBvbmVudCB7XG4gICAgcmVuZGVyKCkge1xuICAgICAgICByZXR1cm4gdGhpcy5wcm9wcy5iYXNrZXQubGVuZ3RoID4gMCA/IChcbiAgICAgICAgICAgIDxkaXYgY2xhc3NOYW1lPVwiYmFza2V0XCI+XG4gICAgICAgICAgICAgICAgPEJhc2tldExpc3QgLz5cbiAgICAgICAgICAgICAgICA8Q2hlY2tvdXRCdXR0b25zIC8+XG4gICAgICAgICAgICAgICAgPEJhc2tldFNoaXBwaW5nIC8+XG4gICAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgKSA6IChcbiAgICAgICAgICAgICfQktCw0YjQsCDQutC+0YDQt9C40L3QsCDQv9GD0YHRgtCwIDooJ1xuICAgICAgICApXG4gICAgfVxufVxuXG5jb25zdCBtYXBTdGF0ZVRvUHJvcHMgPSBzdGF0ZSA9PiB7XG4gICAgcmV0dXJuIHtcbiAgICAgICAgYmFza2V0OiBzdGF0ZS5pdGVtcy5iYXNrZXRcbiAgICB9XG59XG5cbmV4cG9ydCBkZWZhdWx0IGNvbm5lY3QobWFwU3RhdGVUb1Byb3BzKShCYXNrZXQpXG4iXSwibWFwcGluZ3MiOiI7Ozs7Ozs7O0FBQUE7QUFDQTs7O0FBQUE7QUFDQTtBQUFBO0FBQ0E7OztBQUFBO0FBQ0E7OztBQUFBO0FBQ0E7Ozs7Ozs7Ozs7O0FBQ0E7Ozs7Ozs7Ozs7O0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFDQTtBQUNBO0FBQ0E7QUFIQTtBQVFBOzs7O0FBWEE7QUFDQTtBQWFBO0FBQ0E7QUFDQTtBQURBO0FBR0E7QUFDQTtBQUNBIiwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./src/node_modules/basket/index.js\n");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\nvar _list = __webpack_require__(/*! ./list */ \"./src/node_modules/basket/list.js\");\n\nvar _list2 = _interopRequireDefault(_list);\n\nvar _checkout = __webpack_require__(/*! ./checkout */ \"./src/node_modules/basket/checkout.js\");\n\nvar _checkout2 = _interopRequireDefault(_checkout);\n\nvar _shipping = __webpack_require__(/*! basket/shipping */ \"./src/node_modules/basket/shipping.js\");\n\nvar _shipping2 = _interopRequireDefault(_shipping);\n\n__webpack_require__(/*! ./basket.css */ \"./src/node_modules/basket/basket.css\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Basket = function (_Component) {\n    _inherits(Basket, _Component);\n\n    function Basket() {\n        _classCallCheck(this, Basket);\n\n        var _this = _possibleConstructorReturn(this, (Basket.__proto__ || Object.getPrototypeOf(Basket)).call(this));\n\n        _this.switchPage = function (next) {\n            var curPage = _this.state.page;\n            if (next && curPage < 3) _this.setState({\n                page: curPage + 1\n            });else if (!next && curPage > 1) _this.setState({\n                page: curPage - 1\n            });\n        };\n\n        _this.state = {\n            page: 1\n        };\n        return _this;\n    }\n\n    _createClass(Basket, [{\n        key: 'render',\n        value: function render() {\n            var page = this.state.page;\n\n            return this.props.basket.length > 0 ? _react2.default.createElement(\n                'div',\n                { className: 'basket' },\n                page == 1 ? _react2.default.createElement(_list2.default, null) : page == 2 ? _react2.default.createElement(_shipping2.default, null) : null,\n                _react2.default.createElement(_checkout2.default, { page: page, switchPage: this.switchPage })\n            ) : 'Ваша корзина пуста :(';\n        }\n    }]);\n\n    return Basket;\n}(_react.Component);\n\nvar mapStateToProps = function mapStateToProps(state) {\n    return {\n        basket: state.items.basket\n    };\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps)(Basket);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9pbmRleC5qcy5qcyIsInNvdXJjZXMiOlsid2VicGFjazovLy9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9pbmRleC5qcz85ZWNhIl0sInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBSZWFjdCwgeyBDb21wb25lbnQgfSBmcm9tICdyZWFjdCdcclxuaW1wb3J0IHsgY29ubmVjdCB9IGZyb20gJ3JlYWN0LXJlZHV4J1xyXG5pbXBvcnQgQmFza2V0TGlzdCBmcm9tICcuL2xpc3QnXHJcbmltcG9ydCBDaGVja291dEJ1dHRvbnMgZnJvbSAnLi9jaGVja291dCdcclxuaW1wb3J0IEJhc2tldFNoaXBwaW5nIGZyb20gJ2Jhc2tldC9zaGlwcGluZydcclxuaW1wb3J0ICcuL2Jhc2tldC5jc3MnXHJcblxyXG5jbGFzcyBCYXNrZXQgZXh0ZW5kcyBDb21wb25lbnQge1xyXG4gICAgY29uc3RydWN0b3IoKSB7XHJcbiAgICAgICAgc3VwZXIoKVxyXG4gICAgICAgIHRoaXMuc3RhdGUgPSB7XHJcbiAgICAgICAgICAgIHBhZ2U6IDFcclxuICAgICAgICB9XHJcbiAgICB9XHJcbiAgICBzd2l0Y2hQYWdlID0gbmV4dCA9PiB7XHJcbiAgICAgICAgY29uc3QgY3VyUGFnZSA9IHRoaXMuc3RhdGUucGFnZVxyXG4gICAgICAgIGlmIChuZXh0ICYmIGN1clBhZ2UgPCAzKVxyXG4gICAgICAgICAgICB0aGlzLnNldFN0YXRlKHtcclxuICAgICAgICAgICAgICAgIHBhZ2U6IGN1clBhZ2UgKyAxXHJcbiAgICAgICAgICAgIH0pXHJcbiAgICAgICAgZWxzZSBpZiAoIW5leHQgJiYgY3VyUGFnZSA+IDEpXHJcbiAgICAgICAgICAgIHRoaXMuc2V0U3RhdGUoe1xyXG4gICAgICAgICAgICAgICAgcGFnZTogY3VyUGFnZSAtIDFcclxuICAgICAgICAgICAgfSlcclxuICAgIH1cclxuICAgIHJlbmRlcigpIHtcclxuICAgICAgICBjb25zdCB7IHBhZ2UgfSA9IHRoaXMuc3RhdGVcclxuICAgICAgICByZXR1cm4gdGhpcy5wcm9wcy5iYXNrZXQubGVuZ3RoID4gMCA/IChcclxuICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJiYXNrZXRcIj5cclxuICAgICAgICAgICAgICAgIHtwYWdlID09IDEgPyAoXHJcbiAgICAgICAgICAgICAgICAgICAgPEJhc2tldExpc3QgLz5cclxuICAgICAgICAgICAgICAgICkgOiBwYWdlID09IDIgPyAoXHJcbiAgICAgICAgICAgICAgICAgICAgPEJhc2tldFNoaXBwaW5nIC8+XHJcbiAgICAgICAgICAgICAgICApIDogbnVsbH1cclxuICAgICAgICAgICAgICAgIDxDaGVja291dEJ1dHRvbnMgcGFnZT17cGFnZX0gc3dpdGNoUGFnZT17dGhpcy5zd2l0Y2hQYWdlfSAvPlxyXG4gICAgICAgICAgICA8L2Rpdj5cclxuICAgICAgICApIDogKFxyXG4gICAgICAgICAgICAn0JLQsNGI0LAg0LrQvtGA0LfQuNC90LAg0L/Rg9GB0YLQsCA6KCdcclxuICAgICAgICApXHJcbiAgICB9XHJcbn1cclxuXHJcbmNvbnN0IG1hcFN0YXRlVG9Qcm9wcyA9IHN0YXRlID0+IHtcclxuICAgIHJldHVybiB7XHJcbiAgICAgICAgYmFza2V0OiBzdGF0ZS5pdGVtcy5iYXNrZXRcclxuICAgIH1cclxufVxyXG5cclxuZXhwb3J0IGRlZmF1bHQgY29ubmVjdChtYXBTdGF0ZVRvUHJvcHMpKEJhc2tldClcclxuIl0sIm1hcHBpbmdzIjoiOzs7Ozs7OztBQUFBO0FBQ0E7OztBQUFBO0FBQ0E7QUFBQTtBQUNBOzs7QUFBQTtBQUNBOzs7QUFBQTtBQUNBOzs7QUFBQTtBQUNBOzs7Ozs7Ozs7QUFDQTs7O0FBQ0E7QUFBQTtBQUNBO0FBREE7QUFDQTtBQURBO0FBT0E7QUFDQTtBQUVBO0FBREE7QUFLQTtBQURBO0FBR0E7QUFDQTtBQWZBO0FBQ0E7QUFEQTtBQUZBO0FBS0E7QUFDQTs7O0FBV0E7QUFBQTtBQUNBO0FBQ0E7QUFDQTtBQUFBO0FBQ0E7QUFLQTtBQU5BO0FBV0E7Ozs7OztBQUdBO0FBQ0E7QUFDQTtBQURBO0FBR0E7QUFDQTtBQUNBIiwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./src/node_modules/basket/index.js\n");
 
 /***/ }),
 
@@ -43,7 +65,7 @@ eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar List = function (_Component) {\n    _inherits(List, _Component);\n\n    function List() {\n        _classCallCheck(this, List);\n\n        return _possibleConstructorReturn(this, (List.__proto__ || Object.getPrototypeOf(List)).apply(this, arguments));\n    }\n\n    _createClass(List, [{\n        key: 'render',\n        value: function render() {\n            return _react2.default.createElement(\n                'div',\n                { className: 'basket-list' },\n                this.props.basket.map(function (item) {\n                    return _react2.default.createElement(\n                        'div',\n                        { className: 'basket-item', key: item.item.id },\n                        item.item.name,\n                        ' - ',\n                        item.count\n                    );\n                })\n            );\n        }\n    }]);\n\n    return List;\n}(_react.Component);\n\nvar mapStateToProps = function mapStateToProps(state) {\n    return {\n        basket: state.items.basket\n    };\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps)(List);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9saXN0LmpzLmpzIiwic291cmNlcyI6WyJ3ZWJwYWNrOi8vL3NyYy9ub2RlX21vZHVsZXMvYmFza2V0L2xpc3QuanM/YmYzZiJdLCJzb3VyY2VzQ29udGVudCI6WyJpbXBvcnQgUmVhY3QsIHtDb21wb25lbnR9IGZyb20gJ3JlYWN0J1xuaW1wb3J0IHtjb25uZWN0fSBmcm9tICdyZWFjdC1yZWR1eCdcblxuY2xhc3MgTGlzdCBleHRlbmRzIENvbXBvbmVudCB7XG4gICAgcmVuZGVyKCkge1xuICAgICAgICByZXR1cm4gKFxuICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJiYXNrZXQtbGlzdFwiPlxuICAgICAgICAgICAgICAgIHt0aGlzLnByb3BzLmJhc2tldC5tYXAoaXRlbSA9PiB7XG4gICAgICAgICAgICAgICAgcmV0dXJuIChcbiAgICAgICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT0nYmFza2V0LWl0ZW0nIGtleT17aXRlbS5pdGVtLmlkfT5cbiAgICAgICAgICAgICAgICAgICAge2l0ZW0uaXRlbS5uYW1lfSAtIHtpdGVtLmNvdW50fVxuICAgICAgICAgICAgICAgIDwvZGl2PlxuICAgICAgICAgICAgICAgIClcbiAgICAgICAgICAgIH0pfVxuICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgIClcbiAgICB9XG59XG5jb25zdCBtYXBTdGF0ZVRvUHJvcHMgPSAoc3RhdGUpID0+IHtcbiAgICByZXR1cm4ge1xuICAgICAgICBiYXNrZXQ6IHN0YXRlLml0ZW1zLmJhc2tldFxuICAgIH1cbn1cblxuZXhwb3J0IGRlZmF1bHQgY29ubmVjdChtYXBTdGF0ZVRvUHJvcHMpKExpc3QpXG4iXSwibWFwcGluZ3MiOiI7Ozs7Ozs7O0FBQUE7QUFDQTs7O0FBQUE7QUFDQTs7Ozs7Ozs7O0FBQ0E7Ozs7Ozs7Ozs7O0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFDQTtBQUNBO0FBQ0E7QUFBQTtBQUNBO0FBREE7QUFDQTtBQURBO0FBSUE7QUFQQTtBQVVBOzs7O0FBYkE7QUFDQTtBQWNBO0FBQ0E7QUFDQTtBQURBO0FBR0E7QUFDQTtBQUNBIiwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./src/node_modules/basket/list.js\n");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar List = function (_Component) {\n    _inherits(List, _Component);\n\n    function List() {\n        _classCallCheck(this, List);\n\n        return _possibleConstructorReturn(this, (List.__proto__ || Object.getPrototypeOf(List)).apply(this, arguments));\n    }\n\n    _createClass(List, [{\n        key: 'render',\n        value: function render() {\n            return _react2.default.createElement(\n                'div',\n                { className: 'basket-list' },\n                this.props.basket.map(function (item) {\n                    return _react2.default.createElement(\n                        'div',\n                        { className: 'basket-item', key: item.item.id },\n                        item.item.name,\n                        ' - ',\n                        item.count\n                    );\n                })\n            );\n        }\n    }]);\n\n    return List;\n}(_react.Component);\n\nvar mapStateToProps = function mapStateToProps(state) {\n    return {\n        basket: state.items.basket\n    };\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps)(List);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9saXN0LmpzLmpzIiwic291cmNlcyI6WyJ3ZWJwYWNrOi8vL3NyYy9ub2RlX21vZHVsZXMvYmFza2V0L2xpc3QuanM/YmYzZiJdLCJzb3VyY2VzQ29udGVudCI6WyJpbXBvcnQgUmVhY3QsIHsgQ29tcG9uZW50IH0gZnJvbSAncmVhY3QnXG5pbXBvcnQgeyBjb25uZWN0IH0gZnJvbSAncmVhY3QtcmVkdXgnXG5cbmNsYXNzIExpc3QgZXh0ZW5kcyBDb21wb25lbnQge1xuICAgIHJlbmRlcigpIHtcbiAgICAgICAgcmV0dXJuIChcbiAgICAgICAgICAgIDxkaXYgY2xhc3NOYW1lPVwiYmFza2V0LWxpc3RcIj5cbiAgICAgICAgICAgICAgICB7dGhpcy5wcm9wcy5iYXNrZXQubWFwKGl0ZW0gPT4ge1xuICAgICAgICAgICAgICAgICAgICByZXR1cm4gKFxuICAgICAgICAgICAgICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJiYXNrZXQtaXRlbVwiIGtleT17aXRlbS5pdGVtLmlkfT5cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICB7aXRlbS5pdGVtLm5hbWV9IC0ge2l0ZW0uY291bnR9XG4gICAgICAgICAgICAgICAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgICAgICAgICAgICAgKVxuICAgICAgICAgICAgICAgIH0pfVxuICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgIClcbiAgICB9XG59XG5jb25zdCBtYXBTdGF0ZVRvUHJvcHMgPSBzdGF0ZSA9PiB7XG4gICAgcmV0dXJuIHtcbiAgICAgICAgYmFza2V0OiBzdGF0ZS5pdGVtcy5iYXNrZXRcbiAgICB9XG59XG5cbmV4cG9ydCBkZWZhdWx0IGNvbm5lY3QobWFwU3RhdGVUb1Byb3BzKShMaXN0KVxuIl0sIm1hcHBpbmdzIjoiOzs7Ozs7OztBQUFBO0FBQ0E7OztBQUFBO0FBQ0E7Ozs7Ozs7OztBQUNBOzs7Ozs7Ozs7OztBQUNBO0FBQ0E7QUFDQTtBQUFBO0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFDQTtBQURBO0FBQ0E7QUFEQTtBQUlBO0FBUEE7QUFVQTs7Ozs7O0FBRUE7QUFDQTtBQUNBO0FBREE7QUFHQTtBQUNBO0FBQ0EiLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./src/node_modules/basket/list.js\n");
 
 /***/ }),
 
@@ -55,7 +77,7 @@ eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\nvar _dropdown = __webpack_require__(/*! components/dropdown */ \"./src/node_modules/components/dropdown/index.js\");\n\nvar _dropdown2 = _interopRequireDefault(_dropdown);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar BasketShipping = function (_Component) {\n    _inherits(BasketShipping, _Component);\n\n    function BasketShipping() {\n        _classCallCheck(this, BasketShipping);\n\n        return _possibleConstructorReturn(this, (BasketShipping.__proto__ || Object.getPrototypeOf(BasketShipping)).apply(this, arguments));\n    }\n\n    _createClass(BasketShipping, [{\n        key: 'render',\n        value: function render() {\n            return _react2.default.createElement(\n                'div',\n                { className: 'basket-shipping' },\n                _react2.default.createElement(\n                    'p',\n                    null,\n                    '\\u0410\\u0434\\u0440\\u0435\\u0441 \\u043E\\u0442\\u043F\\u0440\\u0430\\u0432\\u043A\\u0438:'\n                ),\n                _react2.default.createElement('input', { type: 'text' }),\n                _react2.default.createElement(\n                    'p',\n                    null,\n                    '\\u0421\\u043F\\u043E\\u0441\\u043E\\u0431 \\u0434\\u043E\\u0441\\u0442\\u0430\\u0432\\u043A\\u0438:'\n                ),\n                _react2.default.createElement(_dropdown2.default, {\n                    src: this.props.shippingMethods,\n                    title: '\\u0412\\u044B\\u0431\\u0435\\u0440\\u0438\\u0442\\u0435 \\u0441\\u043B\\u0443\\u0436\\u0431\\u0443 \\u0434\\u043E\\u0441\\u0442\\u0430\\u0432\\u043A\\u0438'\n                })\n            );\n        }\n    }]);\n\n    return BasketShipping;\n}(_react.Component);\n\nfunction mapStateToProps(state) {\n    return {\n        shippingMethods: state.items.shippingMethods\n    };\n}\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps)(BasketShipping);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9zaGlwcGluZy5qcy5qcyIsInNvdXJjZXMiOlsid2VicGFjazovLy9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9zaGlwcGluZy5qcz9iNTllIl0sInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBSZWFjdCwge0NvbXBvbmVudH0gZnJvbSAncmVhY3QnXG5pbXBvcnQge2Nvbm5lY3R9IGZyb20gJ3JlYWN0LXJlZHV4J1xuaW1wb3J0IERyb3Bkb3duIGZyb20gJ2NvbXBvbmVudHMvZHJvcGRvd24nXG5cbmNsYXNzIEJhc2tldFNoaXBwaW5nIGV4dGVuZHMgQ29tcG9uZW50IHtcbiAgICByZW5kZXIoKSB7XG4gICAgICAgIHJldHVybiAoXG4gICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cImJhc2tldC1zaGlwcGluZ1wiPlxuICAgICAgICAgICAgICAgIDxwPtCQ0LTRgNC10YEg0L7RgtC/0YDQsNCy0LrQuDo8L3A+XG4gICAgICAgICAgICAgICAgPGlucHV0IHR5cGU9XCJ0ZXh0XCIvPlxuICAgICAgICAgICAgICAgIDxwPtCh0L/QvtGB0L7QsSDQtNC+0YHRgtCw0LLQutC4OjwvcD5cbiAgICAgICAgICAgICAgICA8RHJvcGRvd25cbiAgICAgICAgICAgICAgICAgICAgc3JjPXt0aGlzLnByb3BzLnNoaXBwaW5nTWV0aG9kc31cbiAgICAgICAgICAgICAgICAgICAgdGl0bGU9J9CS0YvQsdC10YDQuNGC0LUg0YHQu9GD0LbQsdGDINC00L7RgdGC0LDQstC60LgnXG4gICAgICAgICAgICAgICAgLz5cbiAgICAgICAgICAgIDwvZGl2PlxuICAgICAgICApXG4gICAgfVxufVxuXG5mdW5jdGlvbiBtYXBTdGF0ZVRvUHJvcHMoc3RhdGUpIHtcbiAgICByZXR1cm4ge1xuICAgICAgICBzaGlwcGluZ01ldGhvZHM6IHN0YXRlLml0ZW1zLnNoaXBwaW5nTWV0aG9kc1xuICAgIH1cbn1cblxuZXhwb3J0IGRlZmF1bHQgY29ubmVjdChcbiAgICBtYXBTdGF0ZVRvUHJvcHMsXG4pKEJhc2tldFNoaXBwaW5nKVxuIl0sIm1hcHBpbmdzIjoiOzs7Ozs7OztBQUFBO0FBQ0E7OztBQUFBO0FBQ0E7QUFBQTtBQUNBOzs7Ozs7Ozs7OztBQUNBOzs7Ozs7Ozs7OztBQUNBO0FBQ0E7QUFDQTtBQUFBO0FBQ0E7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUNBO0FBQ0E7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUNBO0FBQ0E7QUFDQTtBQUZBO0FBSkE7QUFVQTs7OztBQWJBO0FBQ0E7QUFlQTtBQUNBO0FBQ0E7QUFEQTtBQUdBO0FBQ0E7QUFDQSIsInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./src/node_modules/basket/shipping.js\n");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\nvar _dropdown = __webpack_require__(/*! components/dropdown */ \"./src/node_modules/components/dropdown/index.js\");\n\nvar _dropdown2 = _interopRequireDefault(_dropdown);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar BasketShipping = function (_Component) {\n    _inherits(BasketShipping, _Component);\n\n    function BasketShipping() {\n        _classCallCheck(this, BasketShipping);\n\n        return _possibleConstructorReturn(this, (BasketShipping.__proto__ || Object.getPrototypeOf(BasketShipping)).apply(this, arguments));\n    }\n\n    _createClass(BasketShipping, [{\n        key: 'render',\n        value: function render() {\n            return _react2.default.createElement(\n                'div',\n                { className: 'basket-shipping' },\n                _react2.default.createElement(\n                    'p',\n                    { className: 'label' },\n                    '\\u0410\\u0434\\u0440\\u0435\\u0441 \\u043E\\u0442\\u043F\\u0440\\u0430\\u0432\\u043A\\u0438:'\n                ),\n                _react2.default.createElement('input', { type: 'text' }),\n                _react2.default.createElement(\n                    'p',\n                    { className: 'label' },\n                    '\\u0421\\u043F\\u043E\\u0441\\u043E\\u0431 \\u0434\\u043E\\u0441\\u0442\\u0430\\u0432\\u043A\\u0438:'\n                ),\n                _react2.default.createElement(_dropdown2.default, {\n                    src: this.props.shippingMethods,\n                    title: '\\u0412\\u044B\\u0431\\u0435\\u0440\\u0438\\u0442\\u0435 \\u0441\\u043B\\u0443\\u0436\\u0431\\u0443 \\u0434\\u043E\\u0441\\u0442\\u0430\\u0432\\u043A\\u0438'\n                })\n            );\n        }\n    }]);\n\n    return BasketShipping;\n}(_react.Component);\n\nfunction mapStateToProps(state) {\n    return {\n        shippingMethods: state.items.shippingMethods\n    };\n}\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps)(BasketShipping);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9zaGlwcGluZy5qcy5qcyIsInNvdXJjZXMiOlsid2VicGFjazovLy9zcmMvbm9kZV9tb2R1bGVzL2Jhc2tldC9zaGlwcGluZy5qcz9iNTllIl0sInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBSZWFjdCwgeyBDb21wb25lbnQgfSBmcm9tICdyZWFjdCdcclxuaW1wb3J0IHsgY29ubmVjdCB9IGZyb20gJ3JlYWN0LXJlZHV4J1xyXG5pbXBvcnQgRHJvcGRvd24gZnJvbSAnY29tcG9uZW50cy9kcm9wZG93bidcclxuXHJcbmNsYXNzIEJhc2tldFNoaXBwaW5nIGV4dGVuZHMgQ29tcG9uZW50IHtcclxuICAgIHJlbmRlcigpIHtcclxuICAgICAgICByZXR1cm4gKFxyXG4gICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cImJhc2tldC1zaGlwcGluZ1wiPlxyXG4gICAgICAgICAgICAgICAgPHAgY2xhc3NOYW1lPVwibGFiZWxcIj7QkNC00YDQtdGBINC+0YLQv9GA0LDQstC60Lg6PC9wPlxyXG4gICAgICAgICAgICAgICAgPGlucHV0IHR5cGU9XCJ0ZXh0XCIgLz5cclxuICAgICAgICAgICAgICAgIDxwIGNsYXNzTmFtZT1cImxhYmVsXCI+0KHQv9C+0YHQvtCxINC00L7RgdGC0LDQstC60Lg6PC9wPlxyXG4gICAgICAgICAgICAgICAgPERyb3Bkb3duXHJcbiAgICAgICAgICAgICAgICAgICAgc3JjPXt0aGlzLnByb3BzLnNoaXBwaW5nTWV0aG9kc31cclxuICAgICAgICAgICAgICAgICAgICB0aXRsZT1cItCS0YvQsdC10YDQuNGC0LUg0YHQu9GD0LbQsdGDINC00L7RgdGC0LDQstC60LhcIlxyXG4gICAgICAgICAgICAgICAgLz5cclxuICAgICAgICAgICAgPC9kaXY+XHJcbiAgICAgICAgKVxyXG4gICAgfVxyXG59XHJcblxyXG5mdW5jdGlvbiBtYXBTdGF0ZVRvUHJvcHMoc3RhdGUpIHtcclxuICAgIHJldHVybiB7XHJcbiAgICAgICAgc2hpcHBpbmdNZXRob2RzOiBzdGF0ZS5pdGVtcy5zaGlwcGluZ01ldGhvZHNcclxuICAgIH1cclxufVxyXG5cclxuZXhwb3J0IGRlZmF1bHQgY29ubmVjdChtYXBTdGF0ZVRvUHJvcHMpKEJhc2tldFNoaXBwaW5nKVxyXG4iXSwibWFwcGluZ3MiOiI7Ozs7Ozs7O0FBQUE7QUFDQTs7O0FBQUE7QUFDQTtBQUFBO0FBQ0E7Ozs7Ozs7Ozs7O0FBQ0E7Ozs7Ozs7Ozs7O0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFDQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQ0E7QUFDQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQ0E7QUFDQTtBQUNBO0FBRkE7QUFKQTtBQVVBOzs7Ozs7QUFHQTtBQUNBO0FBQ0E7QUFEQTtBQUdBO0FBQ0E7QUFDQSIsInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./src/node_modules/basket/shipping.js\n");
 
 /***/ }),
 
@@ -78,7 +100,7 @@ eval("\nvar content = __webpack_require__(/*! !../../../../node_modules/css-load
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\n__webpack_require__(/*! ./dropdown.css */ \"./src/node_modules/components/dropdown/dropdown.css\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Dropdown = function (_Component) {\n    _inherits(Dropdown, _Component);\n\n    function Dropdown(props) {\n        _classCallCheck(this, Dropdown);\n\n        var _this = _possibleConstructorReturn(this, (Dropdown.__proto__ || Object.getPrototypeOf(Dropdown)).call(this, props));\n\n        _this.handleSelect = function (id) {\n            _this.setState({\n                selected: id,\n                listOpen: !_this.state.listOpen\n            });\n        };\n\n        _this.handleClick = function (e) {\n            if (_this.dd.contains(e.target)) {\n                return;\n            }\n            _this.setState({\n                listOpen: false\n            });\n        };\n\n        _this.toggleList = function () {\n            _this.setState({\n                listOpen: !_this.state.listOpen\n            });\n        };\n\n        _this.state = {\n            list: _this.props.src,\n            selected: 0,\n            listOpen: false\n        };\n        return _this;\n    }\n\n    _createClass(Dropdown, [{\n        key: 'componentWillMount',\n        value: function componentWillMount() {\n            document.addEventListener('mousedown', this.handleClick, false);\n        }\n    }, {\n        key: 'componentWillUnmount',\n        value: function componentWillUnmount() {\n            document.removeEventListener('mousedown', this.handleClick, false);\n        }\n    }, {\n        key: 'render',\n        value: function render() {\n            var _this2 = this;\n\n            var _state = this.state,\n                list = _state.list,\n                selected = _state.selected,\n                listOpen = _state.listOpen;\n            var _props = this.props,\n                title = _props.title,\n                children = _props.children;\n\n            return _react2.default.createElement(\n                'div',\n                { ref: function ref(dd) {\n                        return _this2.dd = dd;\n                    }, className: 'dd-wrapper' },\n                _react2.default.createElement(\n                    'div',\n                    { className: 'dd-header', onClick: function onClick() {\n                            return _this2.toggleList();\n                        } },\n                    _react2.default.createElement(\n                        'div',\n                        { className: 'dd-header-title' },\n                        selected ? list[selected - 1].text : title ? title : children ? children : '...'\n                    )\n                ),\n                listOpen && _react2.default.createElement(\n                    'ul',\n                    { className: 'dd-list' },\n                    list.map(function (item) {\n                        var text = item.text,\n                            id = item.id;\n\n                        return _react2.default.createElement(\n                            'li',\n                            {\n                                className: 'dd-list-item' + (selected ? selected == id ? ' selected' : '' : ''),\n                                key: id,\n                                onClick: function onClick() {\n                                    return _this2.handleSelect(id);\n                                } },\n                            text\n                        );\n                    })\n                )\n            );\n        }\n    }]);\n\n    return Dropdown;\n}(_react.Component);\n\nvar mapStateToProps = function mapStateToProps(state) {\n    return {};\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps)(Dropdown);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2NvbXBvbmVudHMvZHJvcGRvd24vaW5kZXguanMuanMiLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vc3JjL25vZGVfbW9kdWxlcy9jb21wb25lbnRzL2Ryb3Bkb3duL2luZGV4LmpzPzE5MjkiXSwic291cmNlc0NvbnRlbnQiOlsiaW1wb3J0IFJlYWN0LCB7Q29tcG9uZW50fSBmcm9tICdyZWFjdCdcbmltcG9ydCB7Y29ubmVjdH0gZnJvbSAncmVhY3QtcmVkdXgnXG5pbXBvcnQgJy4vZHJvcGRvd24uY3NzJ1xuXG5jbGFzcyBEcm9wZG93biBleHRlbmRzIENvbXBvbmVudCB7XG4gICAgY29uc3RydWN0b3IocHJvcHMpIHtcbiAgICAgICAgc3VwZXIocHJvcHMpXG4gICAgICAgIHRoaXMuc3RhdGUgPSB7XG4gICAgICAgICAgICBsaXN0OiB0aGlzLnByb3BzLnNyYyxcbiAgICAgICAgICAgIHNlbGVjdGVkOiAwLFxuICAgICAgICAgICAgbGlzdE9wZW46IGZhbHNlLFxuICAgICAgICB9XG4gICAgfVxuICAgIGhhbmRsZVNlbGVjdCA9IChpZCkgPT4ge1xuICAgICAgICB0aGlzLnNldFN0YXRlKHtcbiAgICAgICAgICAgIHNlbGVjdGVkOiBpZCxcbiAgICAgICAgICAgIGxpc3RPcGVuOiAhdGhpcy5zdGF0ZS5saXN0T3BlblxuICAgICAgICB9KVxuICAgIH1cbiAgICBoYW5kbGVDbGljayA9IChlKSA9PiB7XG4gICAgICAgIGlmICh0aGlzLmRkLmNvbnRhaW5zKGUudGFyZ2V0KSkge1xuICAgICAgICAgICAgcmV0dXJuO1xuICAgICAgICB9XG4gICAgICAgIHRoaXMuc2V0U3RhdGUoe1xuICAgICAgICAgICAgbGlzdE9wZW46IGZhbHNlXG4gICAgICAgIH0pXG4gICAgfVxuICAgIHRvZ2dsZUxpc3QgPSAoKSA9PiB7XG4gICAgICAgIHRoaXMuc2V0U3RhdGUoe1xuICAgICAgICAgICAgbGlzdE9wZW46ICF0aGlzLnN0YXRlLmxpc3RPcGVuXG4gICAgICAgIH0pXG4gICAgfVxuICAgIGNvbXBvbmVudFdpbGxNb3VudCgpIHtcbiAgICAgICAgZG9jdW1lbnQuYWRkRXZlbnRMaXN0ZW5lcignbW91c2Vkb3duJyx0aGlzLmhhbmRsZUNsaWNrLGZhbHNlKVxuICAgIH1cbiAgICBjb21wb25lbnRXaWxsVW5tb3VudCgpIHtcbiAgICAgICAgZG9jdW1lbnQucmVtb3ZlRXZlbnRMaXN0ZW5lcignbW91c2Vkb3duJyx0aGlzLmhhbmRsZUNsaWNrLGZhbHNlKVxuICAgIH1cbiAgICByZW5kZXIoKSB7XG4gICAgICAgIGNvbnN0IHtsaXN0LCBzZWxlY3RlZCwgbGlzdE9wZW59ID0gdGhpcy5zdGF0ZVxuICAgICAgICBjb25zdCB7dGl0bGUsIGNoaWxkcmVufSA9IHRoaXMucHJvcHNcbiAgICAgICAgcmV0dXJuIChcbiAgICAgICAgICAgIDxkaXYgcmVmPXtkZCA9PiB0aGlzLmRkID0gZGR9IGNsYXNzTmFtZT1cImRkLXdyYXBwZXJcIj5cbiAgICAgICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cImRkLWhlYWRlclwiIG9uQ2xpY2s9eygpID0+IHRoaXMudG9nZ2xlTGlzdCgpfT5cbiAgICAgICAgICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJkZC1oZWFkZXItdGl0bGVcIj5cbiAgICAgICAgICAgICAgICAgICAgICAgIHtcbiAgICAgICAgICAgICAgICAgICAgICAgICAgICBzZWxlY3RlZD9cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgbGlzdFtzZWxlY3RlZC0xXS50ZXh0IDogdGl0bGU/XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIHRpdGxlIDogY2hpbGRyZW4/XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICBjaGlsZHJlbjogJy4uLidcbiAgICAgICAgICAgICAgICAgICAgICAgIH1cbiAgICAgICAgICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgICAgICAgICAge2xpc3RPcGVuICYmIDx1bCBjbGFzc05hbWU9XCJkZC1saXN0XCI+XG4gICAgICAgICAgICAgICAgICAgIHtsaXN0Lm1hcChpdGVtID0+IHtcbiAgICAgICAgICAgICAgICAgICAgICAgIGxldCB7dGV4dCwgaWR9ID0gaXRlbVxuICAgICAgICAgICAgICAgICAgICAgICAgcmV0dXJuIChcbiAgICAgICAgICAgICAgICAgICAgICAgICAgICA8bGlcbiAgICAgICAgICAgICAgICAgICAgICAgICAgICBjbGFzc05hbWU9e2BkZC1saXN0LWl0ZW0ke3NlbGVjdGVkP1xuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICBzZWxlY3RlZD09aWQ/XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAnIHNlbGVjdGVkJzonJ1xuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICA6Jyd9YH0gXG4gICAgICAgICAgICAgICAgICAgICAgICAgICAga2V5PXtpZH0gXG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgb25DbGljaz17KCkgPT4gdGhpcy5oYW5kbGVTZWxlY3QoaWQpfT5cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAge3RleHR9XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPC9saT5cbiAgICAgICAgICAgICAgICAgICAgICAgIClcbiAgICAgICAgICAgICAgICAgICAgfSl9XG4gICAgICAgICAgICAgICAgPC91bD59XG5cbiAgICAgICAgICAgIDwvZGl2PlxuICAgICAgICApXG4gICAgfVxufVxuXG5jb25zdCBtYXBTdGF0ZVRvUHJvcHMgPSBzdGF0ZSA9PiB7XG4gICAgcmV0dXJuIHt9XG59XG5cbmV4cG9ydCBkZWZhdWx0IGNvbm5lY3QoXG4gICAgbWFwU3RhdGVUb1Byb3BzLFxuKShEcm9wZG93bilcbiJdLCJtYXBwaW5ncyI6Ijs7Ozs7Ozs7QUFBQTtBQUNBOzs7QUFBQTtBQUNBO0FBQUE7QUFDQTs7Ozs7Ozs7O0FBQ0E7OztBQUNBO0FBQUE7QUFDQTtBQURBO0FBQ0E7QUFEQTtBQVNBO0FBQ0E7QUFDQTtBQUZBO0FBSUE7QUFDQTtBQWRBO0FBZUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQURBO0FBR0E7QUFDQTtBQXRCQTtBQXVCQTtBQUNBO0FBREE7QUFHQTtBQUNBO0FBekJBO0FBQ0E7QUFDQTtBQUNBO0FBSEE7QUFGQTtBQU9BO0FBQ0E7OztBQW1CQTtBQUNBO0FBQ0E7OztBQUNBO0FBQ0E7QUFDQTs7O0FBQ0E7QUFBQTtBQUNBO0FBREE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFDQTtBQUVBO0FBQ0E7QUFBQTtBQUFBO0FBQUE7QUFDQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQ0E7QUFBQTtBQUFBO0FBRUE7QUFGQTtBQURBO0FBVUE7QUFBQTtBQUFBO0FBQ0E7QUFBQTtBQUFBO0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFDQTtBQUlBO0FBQ0E7QUFBQTtBQUFBO0FBQ0E7QUFQQTtBQVVBO0FBZEE7QUFYQTtBQThCQTs7OztBQXBFQTtBQUNBO0FBc0VBO0FBQ0E7QUFDQTtBQUNBO0FBQ0EiLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./src/node_modules/components/dropdown/index.js\n");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n\n__webpack_require__(/*! ./dropdown.css */ \"./src/node_modules/components/dropdown/dropdown.css\");\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Dropdown = function (_Component) {\n    _inherits(Dropdown, _Component);\n\n    function Dropdown(props) {\n        _classCallCheck(this, Dropdown);\n\n        var _this = _possibleConstructorReturn(this, (Dropdown.__proto__ || Object.getPrototypeOf(Dropdown)).call(this, props));\n\n        _this.handleSelect = function (id) {\n            _this.setState({\n                selected: id,\n                listOpen: !_this.state.listOpen\n            });\n        };\n\n        _this.handleClick = function (e) {\n            if (_this.dd.contains(e.target)) {\n                return;\n            }\n            _this.setState({\n                listOpen: false\n            });\n        };\n\n        _this.toggleList = function () {\n            _this.setState({\n                listOpen: !_this.state.listOpen\n            });\n        };\n\n        _this.state = {\n            list: _this.props.src,\n            selected: 0,\n            listOpen: false\n        };\n        return _this;\n    }\n\n    _createClass(Dropdown, [{\n        key: 'componentWillMount',\n        value: function componentWillMount() {\n            document.addEventListener('mousedown', this.handleClick, false);\n        }\n    }, {\n        key: 'componentWillUnmount',\n        value: function componentWillUnmount() {\n            document.removeEventListener('mousedown', this.handleClick, false);\n        }\n    }, {\n        key: 'render',\n        value: function render() {\n            var _this2 = this;\n\n            var _state = this.state,\n                list = _state.list,\n                selected = _state.selected,\n                listOpen = _state.listOpen;\n            var _props = this.props,\n                title = _props.title,\n                children = _props.children;\n\n            return _react2.default.createElement(\n                'div',\n                { ref: function ref(dd) {\n                        return _this2.dd = dd;\n                    }, className: 'dd-wrapper' },\n                _react2.default.createElement(\n                    'div',\n                    { className: 'dd-header', onClick: function onClick() {\n                            return _this2.toggleList();\n                        } },\n                    _react2.default.createElement(\n                        'div',\n                        { className: 'dd-header-title' },\n                        selected ? list[selected - 1].text : title ? title : children ? children : '...'\n                    )\n                ),\n                listOpen && _react2.default.createElement(\n                    'ul',\n                    { className: 'dd-list' },\n                    list.map(function (item) {\n                        var text = item.text,\n                            id = item.id;\n\n                        return _react2.default.createElement(\n                            'li',\n                            {\n                                className: 'dd-list-item' + (selected ? selected == id ? ' selected' : '' : ''),\n                                key: id,\n                                onClick: function onClick() {\n                                    return _this2.handleSelect(id);\n                                } },\n                            text\n                        );\n                    })\n                )\n            );\n        }\n    }]);\n\n    return Dropdown;\n}(_react.Component);\n\nvar mapStateToProps = function mapStateToProps(state) {\n    return {};\n};\n\nexports.default = (0, _reactRedux.connect)(mapStateToProps)(Dropdown);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9zcmMvbm9kZV9tb2R1bGVzL2NvbXBvbmVudHMvZHJvcGRvd24vaW5kZXguanMuanMiLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vc3JjL25vZGVfbW9kdWxlcy9jb21wb25lbnRzL2Ryb3Bkb3duL2luZGV4LmpzPzE5MjkiXSwic291cmNlc0NvbnRlbnQiOlsiaW1wb3J0IFJlYWN0LCB7Q29tcG9uZW50fSBmcm9tICdyZWFjdCdcbmltcG9ydCB7Y29ubmVjdH0gZnJvbSAncmVhY3QtcmVkdXgnXG5pbXBvcnQgJy4vZHJvcGRvd24uY3NzJ1xuXG5jbGFzcyBEcm9wZG93biBleHRlbmRzIENvbXBvbmVudCB7XG4gICAgY29uc3RydWN0b3IocHJvcHMpIHtcbiAgICAgICAgc3VwZXIocHJvcHMpXG4gICAgICAgIHRoaXMuc3RhdGUgPSB7XG4gICAgICAgICAgICBsaXN0OiB0aGlzLnByb3BzLnNyYyxcbiAgICAgICAgICAgIHNlbGVjdGVkOiAwLFxuICAgICAgICAgICAgbGlzdE9wZW46IGZhbHNlLFxuICAgICAgICB9XG4gICAgfVxuICAgIGhhbmRsZVNlbGVjdCA9IChpZCkgPT4ge1xuICAgICAgICB0aGlzLnNldFN0YXRlKHtcbiAgICAgICAgICAgIHNlbGVjdGVkOiBpZCxcbiAgICAgICAgICAgIGxpc3RPcGVuOiAhdGhpcy5zdGF0ZS5saXN0T3BlblxuICAgICAgICB9KVxuICAgIH1cbiAgICBoYW5kbGVDbGljayA9IChlKSA9PiB7XG4gICAgICAgIGlmICh0aGlzLmRkLmNvbnRhaW5zKGUudGFyZ2V0KSkge1xuICAgICAgICAgICAgcmV0dXJuO1xuICAgICAgICB9XG4gICAgICAgIHRoaXMuc2V0U3RhdGUoe1xuICAgICAgICAgICAgbGlzdE9wZW46IGZhbHNlXG4gICAgICAgIH0pXG4gICAgfVxuICAgIHRvZ2dsZUxpc3QgPSAoKSA9PiB7XG4gICAgICAgIHRoaXMuc2V0U3RhdGUoe1xuICAgICAgICAgICAgbGlzdE9wZW46ICF0aGlzLnN0YXRlLmxpc3RPcGVuXG4gICAgICAgIH0pXG4gICAgfVxuICAgIGNvbXBvbmVudFdpbGxNb3VudCgpIHtcbiAgICAgICAgZG9jdW1lbnQuYWRkRXZlbnRMaXN0ZW5lcignbW91c2Vkb3duJyx0aGlzLmhhbmRsZUNsaWNrLGZhbHNlKVxuICAgIH1cbiAgICBjb21wb25lbnRXaWxsVW5tb3VudCgpIHtcbiAgICAgICAgZG9jdW1lbnQucmVtb3ZlRXZlbnRMaXN0ZW5lcignbW91c2Vkb3duJyx0aGlzLmhhbmRsZUNsaWNrLGZhbHNlKVxuICAgIH1cbiAgICByZW5kZXIoKSB7XG4gICAgICAgIGNvbnN0IHtsaXN0LCBzZWxlY3RlZCwgbGlzdE9wZW59ID0gdGhpcy5zdGF0ZVxuICAgICAgICBjb25zdCB7dGl0bGUsIGNoaWxkcmVufSA9IHRoaXMucHJvcHNcbiAgICAgICAgcmV0dXJuIChcbiAgICAgICAgICAgIDxkaXYgcmVmPXtkZCA9PiB0aGlzLmRkID0gZGR9IGNsYXNzTmFtZT1cImRkLXdyYXBwZXJcIj5cbiAgICAgICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cImRkLWhlYWRlclwiIG9uQ2xpY2s9eygpID0+IHRoaXMudG9nZ2xlTGlzdCgpfT5cbiAgICAgICAgICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJkZC1oZWFkZXItdGl0bGVcIj5cbiAgICAgICAgICAgICAgICAgICAgICAgIHtcbiAgICAgICAgICAgICAgICAgICAgICAgICAgICBzZWxlY3RlZD9cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgbGlzdFtzZWxlY3RlZC0xXS50ZXh0IDogdGl0bGU/XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIHRpdGxlIDogY2hpbGRyZW4/XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICBjaGlsZHJlbjogJy4uLidcbiAgICAgICAgICAgICAgICAgICAgICAgIH1cbiAgICAgICAgICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgICAgICAgICAge2xpc3RPcGVuICYmIDx1bCBjbGFzc05hbWU9XCJkZC1saXN0XCI+XG4gICAgICAgICAgICAgICAgICAgIHtsaXN0Lm1hcChpdGVtID0+IHtcbiAgICAgICAgICAgICAgICAgICAgICAgIGxldCB7dGV4dCwgaWR9ID0gaXRlbVxuICAgICAgICAgICAgICAgICAgICAgICAgcmV0dXJuIChcbiAgICAgICAgICAgICAgICAgICAgICAgICAgICA8bGlcbiAgICAgICAgICAgICAgICAgICAgICAgICAgICBjbGFzc05hbWU9e2BkZC1saXN0LWl0ZW0ke3NlbGVjdGVkP1xuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICBzZWxlY3RlZD09aWQ/XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAnIHNlbGVjdGVkJzonJ1xuICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICA6Jyd9YH0gXG4gICAgICAgICAgICAgICAgICAgICAgICAgICAga2V5PXtpZH0gXG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgb25DbGljaz17KCkgPT4gdGhpcy5oYW5kbGVTZWxlY3QoaWQpfT5cbiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAge3RleHR9XG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgPC9saT5cbiAgICAgICAgICAgICAgICAgICAgICAgIClcbiAgICAgICAgICAgICAgICAgICAgfSl9XG4gICAgICAgICAgICAgICAgPC91bD59XG5cbiAgICAgICAgICAgIDwvZGl2PlxuICAgICAgICApXG4gICAgfVxufVxuXG5jb25zdCBtYXBTdGF0ZVRvUHJvcHMgPSBzdGF0ZSA9PiB7XG4gICAgcmV0dXJuIHt9XG59XG5cbmV4cG9ydCBkZWZhdWx0IGNvbm5lY3QoXG4gICAgbWFwU3RhdGVUb1Byb3BzLFxuKShEcm9wZG93bilcbiJdLCJtYXBwaW5ncyI6Ijs7Ozs7Ozs7QUFBQTtBQUNBOzs7QUFBQTtBQUNBO0FBQUE7QUFDQTs7Ozs7Ozs7O0FBQ0E7OztBQUNBO0FBQUE7QUFDQTtBQURBO0FBQ0E7QUFEQTtBQVNBO0FBQ0E7QUFDQTtBQUZBO0FBSUE7QUFDQTtBQWRBO0FBZUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQURBO0FBR0E7QUFDQTtBQXRCQTtBQXVCQTtBQUNBO0FBREE7QUFHQTtBQUNBO0FBekJBO0FBQ0E7QUFDQTtBQUNBO0FBSEE7QUFGQTtBQU9BO0FBQ0E7OztBQW1CQTtBQUNBO0FBQ0E7OztBQUNBO0FBQ0E7QUFDQTs7O0FBQ0E7QUFBQTtBQUNBO0FBREE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFDQTtBQUVBO0FBQ0E7QUFBQTtBQUFBO0FBQUE7QUFDQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQ0E7QUFBQTtBQUFBO0FBRUE7QUFGQTtBQURBO0FBVUE7QUFBQTtBQUFBO0FBQ0E7QUFBQTtBQUFBO0FBQ0E7QUFDQTtBQUNBO0FBQUE7QUFDQTtBQUlBO0FBQ0E7QUFBQTtBQUFBO0FBQ0E7QUFQQTtBQVVBO0FBZEE7QUFYQTtBQThCQTs7Ozs7O0FBR0E7QUFDQTtBQUNBO0FBQ0E7QUFDQSIsInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./src/node_modules/components/dropdown/index.js\n");
 
 /***/ })
 
